% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.Track.R
\name{aggregate.Track}
\alias{aggregate.Track}
\title{Aggregates Tracks}
\usage{
aggregate.Track(x, by = list(strftime(x@time, format = "\%Y-\%m-\%d")),
  FUN = list(coords = median, ` ` = function(x) x[1]), ...)
}
\arguments{
\item{x}{An object of class \code{\link[trajectories:Track-class]{Track}}.}

\item{by}{A vector with grouping values for each data value in \code{x}.
This can be for example the day of each given \code{POSIXct} value in
order to aggregate to daily values.}

\item{FUN}{A list with functions that define how to aggregate the coordinates
(longitude, latitude) and all other variables contained in the \code{data}
slot of \code{x}. The first list element must have the name \code{"coords"} and
contains the function used in order to aggregate the longitude and latitude values.
This is the median by default. The following element(s) must either be given for each
variable or it must be given only one additional element
with name \code{" "} if one function whould be applied to all
variables in the \code{data} slot of \code{x}.}

\item{...}{Further arguments, will be ignored.}

\item{crs}{A character string describing a projection and datum
in the \code{PROJ.4} format (see \code{\link[rgdal]{projInfo}}).}
}
\value{
The temporally aggregated \code{x}.
}
\description{
\code{aggregate.trackvisits} is the aggregation function for objects of class
\code{\link[trajectories:Track-class]{Track}}. Aggregation means that values
of a \code{\link[trajectories:Track-class]{Track}} object are aggregated
temporally, e.g. from hourly to daily values. Aggregation of the corresponding
data values can be performed by user-defined function for each variable.
}
\examples{
#
}
\seealso{
\code{\link{reorganizeTracks}}, \code{\link{redefineIndices}},
\code{\link{fillGapTrack}}, \code{\link{fillGapTracks}},
\code{\link{extractClustersBuffer}}.
}
