% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateDailyLocationsTrack.R
\name{aggregateDailyLocationsTrack}
\alias{aggregateDailyLocationsTrack}
\title{Aggregates a \code{\link[trajectories:Track-class]{Track}} object to Daily Resolution and to Locations.}
\usage{
aggregateDailyLocationsTrack(currenttrack, crs)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories:Track-class]{Track}} object.}

\item{crs}{A character string describing a projection and datum
in the \code{PROJ.4} format (see \code{\link[rgdal]{projInfo}}).}
}
\value{
An object of class \code{\link[trajectories:Track-class]{Track}} that represents the
input \code{\link[trajectories:Track-class]{Track}} object (\code{currenttrack}) that is aggregated
to a daily resolution and for which longitude, latitude and altitude data represent
centroids for the location that has the main data values for that day that were assigned
to a location.
}
\description{
\code{aggregateDailyLocationsTrack} aggregates the data values
of a \code{\link[trajectories:Track-class]{Track}} object that has been
processed with \code{\link{locationsTrack}} to daily resolution.
Additionally, the funciton aggregates longitude, latitude and
altitude values to locations that were identified with
\code{\link{locationsTrack}} and determines the proportion of
gap values for each day with respect to the time frame defined
in \code{attributes(currenttrack)$night} that was set during
performing \code{\link{locationsTrack}}. The aggregation prioritises
campsites
over short-term visits. This means that if during one day, there are
values for a campsite and a short-term visit, values will be
aggregated only for the campsite. If for a day there are vales for
more than one campsite, the function considerd the campsite with most
data values.
}
\details{
The function does not work with \code{\link{locationsTrack}} objects
with a temporal resolution greater than sub-daily resolution.
}
\examples{
#
}
\seealso{
\code{\link{reorganizeTrack}}.
}
