% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateTrack.R
\name{aggregateTrack}
\alias{aggregateTrack}
\title{Aggregates Tracks}
\usage{
aggregateTrack(x, by = list(strftime(x@time, format = "\%Y-\%m-\%d")))
}
\arguments{
\item{x}{An object of class \code{\link[trajectories:Track-class]{Track}} that
has been processed with \code{\link{reorganizeTracks}}, \code{\link{fillGapTracksNight}}
and \code{\link{locationsTrack}}, so that it contains the variables \code{location},
\code{group}, \code{altitude}, \code{gap}, \code{campsite}, \code{norepeatedvisits},
\code{start} and \code{end}. All other variables will be ignored.}

\item{by}{A vector with grouping values for each data value in \code{x}
corresponding to the time. This can be for example the day of each given
\code{POSIXct} value in order to aggregate to daily values.}
}
\value{
The temporally aggregated \code{x}.
}
\description{
\code{aggregate.trackvisits} temporally aggregates objects of class
\code{\link[trajectories:Track-class]{Track}}. Aggregation means that values
of a \code{\link[trajectories:Track-class]{Track}} object are aggregated
temporally, e.g. from hourly to daily values. The function assumes that
certain variables are present for the \code{\link[trajectories:Track-class]{Track}}
object, specifically \code{location}, \code{group}, \code{altitude}, \code{gap},
\code{campsite}, \code{norepeatedvisits}, \code{start} and \code{end} as
defined in \code{\link{locationsTracks}}.
}
\examples{
#
}
\seealso{
\code{\link{reorganizeTracks}}, \code{\link{redefineIndices}},
\code{\link{fillGapTrack}}, \code{\link{fillGapTracks}},
\code{\link{extractClustersBuffer}}.
}
