% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altitudeDifferenceTrack.R
\name{altitudeDifferenceTrack}
\alias{altitudeDifferenceTrack}
\title{Computes the altitude difference between the highest and lowest point of a Track object.}
\usage{
altitudeDifferenceTrack(currenttrack)
}
\arguments{
\item{currenttrack}{An object of class \code{\link[trajectories:Track]{Track}}
with a variable \code{altitude} in the data slot.}
}
\value{
a numeric value representing the altitudinal difference between the
highest and the lowest point in \code{currenttrack}.
}
\description{
\code{altitudeDifferenceTrack} computes the altitudinal difference between
the highest and the lowest point of a
\code{\link[trajectories:Track]{Track}} object with a variable
\code{altitude} in the data slot. \code{NA}s are removed.
}
\examples{
#
}
\seealso{
.
}
