% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altitudeRangeTrack.R
\name{altitudeRangeTrack}
\alias{altitudeRangeTrack}
\title{Extracts the altitude range from a Track object.}
\usage{
altitudeRangeTrack(currenttrack)
}
\arguments{
\item{currenttrack}{An object of class \code{\link[trajectories:Track]{Track}}
with a variable \code{altitude} in the data slot.}
}
\value{
a numeric vector of length two with the range of the altitude values
of \code{currenttrack}.
}
\description{
\code{altitudeRangeTrack} extracts the altitude range
\code{\link[trajectories:Track]{Track}} object with a variable
\code{altitude} in the data slot. \code{NA}s are removed.
}
\examples{
#
}
\seealso{
.
}
