% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.Track.R
\name{altitudinaldistancenextvisit_wos}
\alias{altitudinaldistancenextvisit_wos}
\title{Computes Altitudinal Move Distances Not Ignoring Short-Term Visits}
\usage{
altitudinaldistancenextvisit_wos(agg, aggcoords)
}
\arguments{
\item{agg}{A \code{data.frame} object as created internally in \code{\link{aggregateTrack}}.}
}
\value{
A numeric vector containing the linear altitudinal distance between locations of adjacent campsites,
whereby short-term visits are counted as gaps. Values for a following
move
are always assigned to the last value of a visit in \code{agg} as indicated by \code{agg$end}.
}
\description{
\code{altitudinaldistancenextvisit_wos} computes for a \code{data.frame} as created
internally in \code{\link{aggregateTrack}} the linear altitudinal distance between locations
of adjacent campsites, whereby short-term visits are counted as gaps.
}
\examples{
#
}
\seealso{
\code{\link{aggregateTrack}}
}
\keyword{internal}
