% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignFixedTenDayInterval.R
\name{assignFixedTenDayInterval}
\alias{assignFixedTenDayInterval}
\title{Assigns days to fixed ten-day intervals.}
\usage{
assignFixedTenDayInterval(timedate, startnew = TRUE)
}
\arguments{
\item{timedate}{A \code{POSIXct} vector containing date (and time)
information.}

\item{startnew}{A logical value indicating if the ID for ten-day intervals
should be reset to \code{1} for each new year (\code{startbew = TRUE}) or
not (\code{startnew = FALSE}).}
}
\value{
An integer vector with the same length as \code{timedate} assigning
to each value of \code{timedate} an ID of a fixed ten-day interval.
}
\description{
\code{assignFixedTenDayInterval} assigns each value of a \code{POSIXct}
vector to a specific fixed ten-day interval of a year (three intervals
per month from day one to 10, 11 to 20 and 21 to 28/29/30/31 for each
month, depending on the month and year. Leap years will be considered).
This is done by constructing an integer vector containing values for
each value of the \code{POSIXct} vector and unique values for each ten-
day interval.
}
\details{
It can be specified if the ID for the ten-day intervals should be reset
for each new year (\code{startnew = TRUE}) or not (\code{startnew = FALSE}).
}
\examples{
#
}
\seealso{
\code{\link{extractRasterTrack}},
\code{\link{extractRasterTracks}}.
}
