% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageDistanceCampsiteLocationsTrack.R
\name{averageDistanceCampsiteLocationsTrack}
\alias{averageDistanceCampsiteLocationsTrack}
\title{Computes the mean distance between campsite locations of a Track object in the order of movement.}
\usage{
averageDistanceCampsiteLocationsTrack(currenttrack, fun = stats::median)
}
\arguments{
\item{currenttrack}{An object of class \code{\link[trajectories:Track-class]{Track}}
with a variable \code{location} and \code{campsite}}

\item{fun}{One of \code{mean} or \code{median)}, depending on which function should be
used in order to compute average values.}
}
\value{
a numeric value representing the number of unique (campsite) locations
in \code{currenttrack}.
}
\description{
\code{averageDistanceCampsiteLocationsTrack} computes the mean distance between
locations of a \code{\link[trajectories:Track-class]{Track}} in the order of movement.
Several options are available.
}
\examples{
#
}
\seealso{
.
}
