% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chullTrack.R
\name{chullTrack}
\alias{chullTrack}
\title{Extracts the convex hull from a Track object.}
\usage{
chullTrack(currenttrack)
}
\arguments{
\item{currenttrack}{An object of class \code{\link[trajectories:Track-class]{Track}}.}
}
\value{
A \code{\link[sp:SpatialPolygons-class]{SpatialPolygons}} object
with coordinates corresponding to coordinates of points of the convex hull.
}
\description{
\code{chullTrack} extracts the convex hull from a
\code{\link[trajectories:Track-class]{Track}} object and returns it
as \code{\link[sp:SpatialPolygons-class]{SpatialPolygons}} object.
}
\examples{
#
}
\seealso{
.
}
