% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyNightTrack.R
\name{classifyNightTrack}
\alias{classifyNightTrack}
\title{Classifies Data Values of a \code{Tracks} object as Night or Day Values.}
\usage{
classifyNightTrack(currenttrack, night)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories:Track-class]{Track}} object.}

\item{night}{An integer vector with two elements:
\enumerate{
  \item The first element specifies the start hour of the night, e.g. \code{0}
  for 0 o'clock.
  \item The first element specifies the start hour of the night, e.g. \code{4}
  for 4 o'clock.
}}
}
\value{
A \code{\link[trajectories:Track-class]{Track}} object identical to \code{currenttrack}
with an attribute \code{night} representing a logical vector with a value for
each data value in \code{currenttrack} that is \code{TRUE} if a data value was
recorded during night and \code{FALSE} if it was recorded during day.
}
\description{
\code{classifyNightTrack} creates labels for the data values
of a \code{\link[trajectories:Track-class]{Track}} object marking
a data value as recorded during night or during day. The decision
boundaries are specified by the user.
}
\examples{
#
}
\seealso{
#
}
