% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterOrder.R
\name{clusterOrder}
\alias{clusterOrder}
\title{Identifies clusters in GPS tracks considered as campsites.}
\usage{
clusterOrder(currenttracks, cellsize = 800, threshold = 20)
}
\arguments{
\item{currenttracks}{A \code{\link[trajectories:Track-class]{Tracks}} object.}

\item{cellsize}{Numerical value representing the target size of the
quadrats to create.}

\item{threshold}{An integer value specifying the minimum number of
points within a grid cell needed in order to consider these points
as cluster.}
}
\value{
A \code{\link[sp:SpatialPoints]{SpatialPointsDataFrame}} holding grid cells
determined as clusters including data on the number of visits per cluster,
start and end times etc.
}
\description{
\code{clusterOrder} identifies campsites (clusters of points) of a household
(\code{\link[trajectories:Track-class]{Tracks}} object) and for each cluster/campsite
determines the number of visits, the order in which the clusters have been
visited, and the start and end times (first and last point) of each visit.
}
\details{
\code{clusterOrder} uses the function \code{\link{extractClusters}} to
detect the campsites. \code{clusterOrder} also uses the functions
\code{link{searchNextVisit}} and \code{\link{findStarts}}.
}
\examples{
#
}
\seealso{
\code{\link{qTopology}}, \code{\link{pointsPerQuad}},
\code{\link{extractClusters}}, \code{\link{searchNextVisit}},
\code{\link{findStarts}}.
}
