% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countGaps.R
\name{countGaps}
\alias{countGaps}
\title{Counts gaps in GPS tracks.}
\usage{
countGaps(currenttrack, tstart = min(index(currenttrack@time)),
  tend = max(index(currenttrack@time)), threshold = 18000,
  tolerance = 2)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories:Track-class]{Track}} object.}

\item{tstart}{A \code{POSIXct} object indicating the date and time of the start
of the target period. Default is the first value of the
\code{\link[trajectories]{Track}} object.}

\item{tend}{A \code{POSIXct} object indicating the date and time of the end of
the target period. Default is the last value of the
\code{\link[trajectories]{Track}} object.}

\item{threshold}{An integer value indicating the threshold for the time between
two data points [s] above which this is considered as a gap in the
\code{\link[trajectories:Track-class]{Track}} object. Default is \code{threshold = 18000},
i.e. 5 hours. If \code{threshold = NULL}, the threshold value is calculated from
the data as \code{mean(duration) + tolerance * sd(duration)}.}

\item{tolerance}{An integer value indicating the tolerance of the threshold value
calculated from the data if \code{threshold = NULL}. Default is
\code{tolerance = 2}, i.e. 2 seconds.}
}
\value{
A numeric value representing the number of gaps within the
\code{\link[trajectories:Track-class]{Track}} object.
}
\description{
\code{countGaps} counts gaps in a given \code{\link[trajectories:Track-class]{Track}} object
(for an individual \code{\link[trajectories:Track-class]{Track}} object) within a given time
period, using the function \code{\link{findGaps}}.
}
\examples{
#
}
\seealso{
\code{\link{findGaps}}, \code{\link{gapDuration}},
\code{\link{gapProportion}}, \code{\link{totalGapProportion}}.
}
