% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countUniqueLocationsTrack.R
\name{countUniqueLocationsTrack}
\alias{countUniqueLocationsTrack}
\title{Computes the number of unique locations of a Track object.}
\usage{
countUniqueLocationsTrack(currenttrack, campsite = TRUE,
  repeated = FALSE)
}
\arguments{
\item{currenttrack}{An object of class \code{\link[trajectories:Track]{Track}}
with a variable \code{location}, \code{campsite} and \code{norepeatedcampsitevisits}.
as created by \code{\link{locationsTrack}} in the data slot.}

\item{campsite}{A logical value indicating if only campsites should be considered
(\code{TRUE}) or any locations (\code{FALSE}).}

\item{repeated}{A logical value indicating if the number of locations with repeated
visits shoud be counted (\code{TRUE}) or not (\code{FALSE}). This is only available
with \code{campsite = TRUE}.}
}
\value{
a numeric value representing the number of unique (campsite) locations
in \code{currenttrack}.
}
\description{
\code{countUniqueLocationsTrack} counts unique lcoations of a
\code{\link[trajectories:Track]{Track}}. Several options are available.
}
\examples{
#
}
\seealso{
.
}
