% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daynightFixesTrack.R
\name{daynightFixesTrack}
\alias{daynightFixesTrack}
\title{Classifies sub-daily values of GPS tracks.}
\usage{
daynightFixesTrack(currenttrack, tstart = 16, tend = 20)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories]{Track}} object.}

\item{tstart}{A numeric value indicating the start of the time interval
[h, values between 0 and 24]. \code{tstart} has to be smaller than
\code{tend}.}

\item{tend}{A numeric value indicating the end of the time interval
[h, values between 0 and 24]. \code{tend} has to be larger than
\code{tstart}.}
}
\value{
a \code{\link[sp]{SpatialPointsDataFrame}} containing all data
of the slot \code{data} of the input \code{\link[trajectories]{Track}}
object and a column \code{night} indicating if a data point is within
the specified time interval (\code{night = 1}) or not
(\code{night = 0}).
}
\description{
\code{daynightFixesTrack} classifies the values of a
\code{\link[trajectories]{Track}} object in two classes on a sub-daily basis
by defining threshold hour values (for example in order to classify values as
recorded during daytime and nighttime).
}
\examples{
#
}
\seealso{
\code{\link{downByDrop}}, \code{\link{downByDropTrs}},
\code{\link{getNightTrack}}, \code{\link{getNightTrs}},
\code{\link{dayNightFixesTracks}},
\code{\link{getNightConnectionsTrack}}.
}
