% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateTrack.R
\name{distancenextvisit_wis}
\alias{distancenextvisit_wis}
\title{Computes Move Distances Not Ignoring Short-Term Visits}
\usage{
distancenextvisit_wis(agg, aggcoords)
}
\arguments{
\item{agg}{A \code{data.frame} object as created internally in \code{\link{aggregateTrack}}.}

\item{aggcoords}{A \code{\link[sp:SpatialPoints-class]{SpatialPoints}} object with a value for each value
in \code{agg}.}
}
\value{
A numeric vector containing the linear distance between locations of adjacent visits,
whereby short-term visits are not counted as gaps and are considered as visits. Values for a following
move
are always assigned to the last value of a visit in \code{agg} as indicated by \code{agg$end}.
}
\description{
\code{distancenextvisit_wis} computes for a \code{data.frame} as created
internally in \code{\link{aggregateTrack}} the linear distance between locations
of adjacent visits, whereby short-term visits are not counted as gaps and
are considered as visits.
}
\examples{
#
}
\seealso{
\code{\link{aggregateTrack}}
}
\keyword{internal}
