% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateTrack.R
\name{distancenextvisit_wis_campsite}
\alias{distancenextvisit_wis_campsite}
\title{Computes Move Distances Not Ignoring Short-Term Visits}
\usage{
distancenextvisit_wis_campsite(agg)
}
\arguments{
\item{agg}{A \code{data.frame} object as created internally in \code{\link{aggregateTrack}}.}
}
\value{
A numeric vector containing the linear distance between locations
of adjacent campsites along locations of short-term visits (if present). Values for a following
move
are always assigned to the last value of a campsite in \code{agg} as indicated by \code{agg$end}.
}
\description{
\code{distancenextvisit_wis_campsite} computes for a \code{data.frame} as created
internally in \code{\link{aggregateTrack}} the linear distance between locations
of adjacent campsites along locations of short-term visits (if present). This is done
by summing all distances covered between adjacent campsites as computed with
\code{\link{distancenextvisit_wis}}.
}
\examples{
#
}
\seealso{
\code{\link{aggregateTrack}}
}
\keyword{internal}
