% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateTrack.R
\name{distancenextvisit_wos}
\alias{distancenextvisit_wos}
\title{Computes Move Distances Not Ignoring Short-Term Visits}
\usage{
distancenextvisit_wos(agg, aggcoords)
}
\arguments{
\item{agg}{A \code{data.frame} object as created internally in \code{\link{aggregateTrack}}.}

\item{aggcoords}{A \code{\link[sp]{SpatialPoints}} object with a value for each value
in \code{agg}.}
}
\value{
A numeric vector containing the linear distance between locations of adjacent campsites,
whereby short-term visits are counted as gaps. Values for a following
move
are always assigned to the last value of a visit in \code{agg} as indicated by \code{agg$end}.
}
\description{
\code{distancenextvisit_wos} computes for a \code{data.frame} as created
internally in \code{\link{aggregateTrack}} the linear distance between locations
of adjacent campsites, whereby short-term visits are counted as gaps.
}
\examples{
#
}
\seealso{
\code{\link{aggregateTrack}}
}
\keyword{internal}
