% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downByDropTrackscollection.R
\name{downByDropTrackscollection}
\alias{downByDropTrackscollection}
\title{Cleans and regularises GPS tracks.}
\usage{
downByDropTrackscollection(currenttrackscollection, timeInterval = 1770,
  cores = 1, clcall = NULL)
}
\arguments{
\item{currenttrackscollection}{A
\code{\link[trajectories:Track-class]{TracksCollection}} object.}

\item{timeInterval}{An integer value indicating a time interval [s] as
target duration between data points of the \code{\link[trajectories:Track-class]{Track}}
object. Default is \code{timeInterval = 1770} (i.e. 29.5 minutes).}

\item{cores}{An integer value representing the number of cores to
use in parallel computing.}

\item{clcall}{A function that is passed to
\code{\link[parallel]{clusterCall}}.}
}
\value{
The downsampled \code{\link[trajectories:Track-class]{TracksCollection}}
object.
}
\description{
\code{downByDropTrackscollection} subsets the values of all
\code{\link[trajectories:Track-class]{Track}} objects of all
\code{\link[trajectories:Track-class]{Tracks}} objects of a given
\code{\link[trajectories:Track-class]{TracksCollection}} object to a regular time
interval by using the function \code{\link{downByDrop}}.
}
\examples{
#
}
\seealso{
\code{\link{downByDrop}}, \code{\link{getNightTrack}},
\code{\link{getNightTrs}}, \code{\link{daynightFixesTrack}},
\code{\link{daynightFixesTracks}},
\code{\link{getNightConnectionsTrack}}.
}
