% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downByDropTrs.R
\name{downByDropTrs}
\alias{downByDropTrs}
\title{Cleans and regularises GPS tracks.}
\usage{
downByDropTrs(currenttracks, timeInterval = 1770)
}
\arguments{
\item{currenttracks}{A \code{\link[trajectories]{Tracks}} object.}

\item{timeInterval}{An integer value indicating a time interval [s] as
target duration between data points of the \code{\link[trajectories]{Track}}
object. Default is \code{timeInterval = 1770} (i.e. 29.5 minutes).}
}
\value{
The downsampled \code{\link[trajectories]{Tracks}} object.
}
\description{
\code{downByDropTrs} subsets the values of all \code{\link[trajectories]{Track}}
objects of a given \code{\link[trajectories]{Tracks}} object to a regular time
interval by using the function \code{\link{downByDrop}}.
}
\examples{
#
}
\seealso{
\code{\link{downByDrop}}, \code{\link{getNightTrack}},
\code{\link{getNightTrs}}, \code{\link{daynightFixesTrack}},
\code{\link{daynightFixesTracks}},
\code{\link{getNightConnectionsTrack}}.
}
