% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exDataTracks.R
\name{exDataTracks}
\alias{exDataTracks}
\title{Extracts data from a \code{\link[trajectories]{Tracks}} object.}
\usage{
exDataTracks(currenttracks)
}
\arguments{
\item{currenttracks}{A \code{\link[trajectories]{Tracks}} object.}
}
\value{
A \code{data.frame} objects representing the merged slots
\code{data} and \code{time} of the \code{\link[trajectories]{Track}}
objects of the specified \code{\link[trajectories]{Tracks}}
object.
}
\description{
\code{exDataTracks} extracts data in the \code{data} slot and
\code{time} slot of all \code{\link[trajectories]{Track}} objects
of specified \code{\link[trajectories]{Tracks}} object and merges
the values into a \code{data.frame} for each
\code{\link[trajectories]{Track}} object
}
\examples{
#
}
