% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractClusters.R
\name{extractClusters}
\alias{extractClusters}
\title{Extracts clusters based on point counts in a spatial grid.}
\usage{
extractClusters(currenttracks, targetQsize = 800, threshold = 20)
}
\arguments{
\item{currenttracks}{A \code{\link[trajectories]{Tracks}} object.}

\item{targetQsize}{Numerical value representing the target size of the
quadrats to create.}

\item{threshold}{An integer value specifying the minimum number of
points within a grid cell needed in order to consider these points
as cluster.}
}
\value{
A \code{\link[sp]{SpatialGridDataFrame}} with the cells
determined as clusters.
}
\description{
\code{extractClusters} extract clusters using the function
\code{\link{pointsPerQuad}}.
}
\examples{
#
}
\seealso{
\code{\link{qTopology}}, \code{\link{pointsPerQuad}},
\code{\link{searchNextVisit}}, \code{\link{findStarts}},
\code{\link{clusterOrder}}.
}
