% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCoordinatesVisitsTrack.R
\name{extractCoordinatesVisitsTrack}
\alias{extractCoordinatesVisitsTrack}
\title{Extracts centroid coordinates of visits from GPS tracks.}
\usage{
extractCoordinatesVisitsTrack(currenttrack, aggregated = FALSE,
  crs = "+proj=utm +zone=46 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0")
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories:Track-class]{Track}} object as
returned by \code{link{locationsTrack}}}

\item{aggregated}{A logical value indicating if centroid coordinates
should be computed for all visits at each location
(i.e. without aggregation of short-term visits with small temporal
interstices) (\code{aggregated = FALSE}) or for short-term visits and
long-term visits (cmapsites) (\code{aggregated = TRUE}) that may be
aggregated from visits at the same location if the interstice bewteen
the visits was smaller than \code{tmaxinterstices} as parametrised in
the previous \code{link{locationsTrack}} call.}
}
\value{
A \code{data.frame} object with centroid longitude values in
the first column, centroid latitude values in the second column and
centroid altitude values in the third column for each visit of a location
(as indicated by unique combinations of \code{currenttrack$location} and
\code{currenttrack$visitsloc}). Gaps will be omitted. The \code{data.frame}
object has the same number of rows as \code{currenttrack}.
}
\description{
\code{centroidCoordinatesVisitsTrack} extract the mean coordinates for
visits at locations in a \code{\link[trajectories:Track-class]{Track}} object as
returned by \code{\link{locationsTrack}}.
}
\details{
Centroid coordinates can be computed for all visits at each location
(i.e. without aggregation of short-term visits with small temporal
interstices) (\code{aggregated = FALSE}) or for short-term visits and
long-term visits (cmapsites) (\code{aggregated = TRUE}) that may be
aggregated from visits at the same location if the interstice bewteen
the visits was smaller than \code{tmaxinterstices} as parametrised in
the previous \code{link{locationsTrack}} call.
}
\examples{
#
}
\seealso{
\code{\link{identifyTimeIntervals}}, \code{\link{removeDataTrack}},
\code{\link{removeDataTracks}}, \code{\link{nogapDurationTracks}}.
}
