% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRasterTrack.R
\name{extractRasterTrack}
\alias{extractRasterTrack}
\title{Extracts corresponding raster values for GPS tracks.}
\usage{
extractRasterTrack(currenttrack, raster, timedate, resolution)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories]{Track}} object that
has a column \code{time} containing the time information of the data
values.}

\item{raster}{raster based time series (\code{RasterBrick} or
\code{RasterStack} object, see: \code{\link[raster]{Raster-class}}).}

\item{timedate}{A \code{POSIXct} vector with values for each layer of
\code{raster}.}

\item{resolution}{A character value indicating the temporal resolution
of \code{raster} and \code{timedate}, respectively. One of \code{"days"},
\code{"fixedtendays"}, \code{"movingwindowtendays"} and \code{"months"}.}
}
\value{
A vector with a value for each data value of \code{currenttrack}.
If a value of \code{raster} cannot be assigned to a respective data value,
\code{NA} is returned for the respective data value.
}
\description{
\code{extractRasterTrack} extracts values from raster based time series
(\code{RasterBrick} or \code{RasterStack} object, see:
\code{\link[raster]{Raster-class}}) that correspond the the respective
position and time of a \code{\link[trajectories]{Track}} object.
}
\examples{
#
}
\seealso{
\code{\link{assignFixedTenDayInterval}},
\code{\link{extractRasterTracks}}.
}
