% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRasterTrack1.R
\name{extractRasterTrack1}
\alias{extractRasterTrack1}
\title{Extracts corresponding raster values for GPS tracks.}
\usage{
extractRasterTrack1(currenttrack, raster, buffer = 0)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories:Track-class]{Track}} object that
has a column \code{day} containing the time information of the data
values.}

\item{raster}{raster based time series (\code{RasterBrick} or
\code{RasterStack} object, see: \code{\link[raster]{Raster-class}}).
Eacj layer of \code{raster} has to correspond to a value in
\code{currenttrack}, i.e. both must have daily resolution and
the same number of values and layers, respectively.}

\item{buffer}{A numeric value indicating the radius of the buffer
around each point that should be considered during extraction of
the raster values. If the data are not projected
(latitude/longitude), the unit should be meters. Otherwise it should
be in map-units (typically also meters).}
}
\value{
A vector with a value for each data value of \code{currenttrack}.
If a value of \code{raster} cannot be assigned to a respective data value,
\code{NA} is returned for the respective data value.
}
\description{
\code{extractRasterTrack1} extracts values from raster based time series
(\code{RasterBrick} or \code{RasterStack} object, see:
\code{\link[raster]{Raster-class}}) that correspond the the respective
position and time of a \code{\link[trajectories:Track-class]{Track}} object with
daily resolution.
}
\examples{
#
}
\seealso{
\code{\link{assignFixedTenDayInterval}},
\code{\link{extractRasterTracks}}.
}
