% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRasterTrackSingleLayer.R
\name{extractRasterTrackSingleLayer}
\alias{extractRasterTrackSingleLayer}
\title{Extracts corresponding raster values for GPS tracks.}
\usage{
extractRasterTrackSingleLayer(currenttrack, raster, buffer = 0)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories:Track-class]{Track}} object that
has a column \code{day} containing the time information of the data
values.}

\item{raster}{A \code{RasterLayer} (see: \code{\link[raster]{Raster-class}}).
from which values should be extracted.}

\item{buffer}{A numeric value indicating the radius of the buffer
around each point that should be considered during extraction of
the raster values. If the data are not projected
(latitude/longitude), the unit should be meters. Otherwise it should
be in map-units (typically also meters).}
}
\value{
A vector with a value for each data value of \code{currenttrack}.
If a value of \code{raster} cannot be assigned to a respective data value,
\code{NA} is returned for the respective data value.
}
\description{
\code{extractRasterTrackSingleLayer} extracts values from \code{RasterLayer}
(see: \code{\link[raster]{Raster-class}}) that corresponds to the the respective
position of a \code{\link[trajectories:Track-class]{Track}} object with
daily resolution.
}
\examples{
#
}
\seealso{
\code{\link{assignFixedTenDayInterval}},
\code{\link{extractRasterTracks}}.
}
