% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRasterTrackTimeinterval.R
\name{extractRasterTrackTimeinterval}
\alias{extractRasterTrackTimeinterval}
\title{Extracts corresponding raster values for GPS tracks.}
\usage{
extractRasterTrackTimeinterval(currenttrack, raster, buffer = 0)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories:Track-class]{Track}} object that
has a column \code{day} containing the time information of the data
values.}

\item{raster}{A \code{RasterBrick} or \code{RasterStack} (see:
\code{\link[raster:Raster-classes]{Raster-classes}}). from which values should be
extracted. Each layer has to correspond to a time interval with
a duration longer than a day.}

\item{buffer}{A numeric value indicating the radius of the buffer
around each point that should be considered during extraction of
the raster values. If the data are not projected
(latitude/longitude), the unit should be meters. Otherwise it should
be in map-units (typically also meters).}

\item{timedate}{A \code{POSIXct} vector with each element representing
the first day of the time interval the layers of \code{raster} correspond
to plus one element representing the first day after the last day of the
last time interval.}
}
\value{
A vector with a value for each data value of \code{currenttrack}.
If a value of \code{raster} cannot be assigned to a respective data value,
\code{NA} is returned for the respective data value.
}
\description{
\code{extractRasterTrackTimeinterval} extracts values from a \code{RasterBrick}
or \code{RasterStack} object (see: \code{\link[raster:Raster-classes]{Raster-classes}}) with a
temporal resolution greater than one day that corresponds to the respective
position of values of a \code{\link[trajectories:Track-class]{Track}} object with
daily resolution.
}
\examples{
#
}
\seealso{
\code{\link{assignFixedTenDayInterval}},
\code{\link{extractRasterTracks}}.
}
