% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSpatialPolygonsDataFrameTrack.R
\name{extractSpatialPolygonsDataFrameTrack}
\alias{extractSpatialPolygonsDataFrameTrack}
\title{Assigns \code{\link[sp]{SpatialPolygonsDataFrame}} values to values of a GPS tracks.}
\usage{
extractSpatialPolygonsDataFrameTrack(currenttrack, sppdf, variable)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories]{Track}} object that
has a column \code{time} containing the time information of the data
values.}

\item{sppdf}{A \code{\link[sp]{SpatialPolygonsDataFrame}} object.}

\item{variable}{A \code{character} value indicating for which variable
to extract values from \code{sppdf}.}
}
\value{
A vector with a value for each data value of \code{currenttrack}
extracted from the speified variable of \code{sppdf}.
If a value of \code{raster} cannot be assigned to a respective data value,
\code{NA} is returned for the respective data value.
}
\description{
\code{extractSpatialPolygonsDataFrameTrack} extracts for each data
value of a \code{\link[trajectories]{Track}} object a value
of a \code{\link[sp]{SpatialPolygonsDataFrame}} object of a specified
variable.
}
\examples{
#
}
\seealso{
.
}
