% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillGapTrackNight.R
\name{fillGapTrackNight}
\alias{fillGapTrackNight}
\title{Imputes gaps in a \code{Track} object.}
\usage{
fillGapTrackNight(currenttrack, maxduration, maxdistance, night = c(16, 20))
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories:Track-class]{Track}} object with a
boolean column \code{gap} in \code{currenttrack@data}. Data values
have to be regularly spaced (may be achieved for example with
\code{\link{reorganizeTracks}}).}

\item{maxduration}{A numerical value representing the maximum
allowed duration of a gap that is filled [s]. It has to be paid
attention to the fact that \code{maxduration} should be adapted to
the time interval specified by \code{night}. See the details section.}

\item{maxdistance}{A numerical value representing the maximum
allowed distance between the spatial position of the last data
value before a gap and the spatial position of the first data
value after a gap that is filled [m].}

\item{night}{An integer vector with two elements:
\enumerate{
  \item The first element specifies the start hour of the night, e.g. \code{0}
  for 0 o'clock.
  \item The first element specifies the start hour of the night, e.g. \code{4}
  for 4 o'clock.
}}
}
\value{
The input \code{\link[trajectories:Track-class]{Track}} object with
the variable \code{gap} set to \code{FALSE} for filled gap values. Additionally,
the output has a new variable \code{filled} indicating if a gap was filled
(\code{filled = TRUE}) or not (\code{filled = FALSE})
}
\description{
\code{fillGapTrackNight} imputes missing values in a
\code{\link[trajectories:Track-class]{Track}} object. Gaps are filled
if their duration is \eqn{\le} a user specified duration threshold
and if the distance between the spatial position of the last data
value before the gap and the spatial position of the first data
value after the gap is \eqn{\le} a user specified distance threshold.
}
\details{
It has to be paid
attention to the fact that \code{maxduration} should be adapted to
the time interval specified by \code{night}. For example, if one specifies
the time interval between 0 and 4 o' clock as night and wants to allow
a \code{maxduration} of four days (assuming that there may be gaps for four
following nights, but if there is no gap in the fifth night, then the gap is
filled), then one must specifiy \code{maxduration} as
\code{4*24*60*60 + (24 - nightduration)*60*60}, whereby \code{nightduration}
is the duration of the time interval specified as night in hours.
}
\examples{
#
}
\seealso{
\code{\link{fillGapTracksNight}}.
}
