% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillGapTracks.R
\name{fillGapTracks}
\alias{fillGapTracks}
\title{Imputes gaps in a \code{\link[trajectories]{Tracks}} object.}
\usage{
fillGapTracks(currenttracks, maxduration, maxdistance, timeinterval,
  cores = 1, clcall = NULL)
}
\arguments{
\item{currenttracks}{A \code{\link[trajectories]{Tracks}} object with
\code{\link[trajectories]{Track}} objects with a
boolean column \code{gap} in \code{track@data}. Data values
have to be regularly spaced (may be achieved for example with
\code{\link{reorganizeTracks}}).}

\item{maxduration}{A numerical value representing the maximum
allowed duration of a gap that is filled [s].}

\item{maxdistance}{A numerical value representing the maximum
allowed distance between the spatial position of the last data
value before a gap and the spatial position of the first data
value after a gap that is filled [m].}

\item{timeinterval}{A numerical value reperesenting the duration
of a time interval represented by one data value of \code{track} [s].}

\item{cores}{An integer value representing the number of cores to
use in parallel computing.}

\item{clcall}{A function that is passed to
\code{\link[parallel]{clusterCall}}.}
}
\value{
The input \code{\link[trajectories]{Tracks}} object with filled
gaps.
}
\description{
\code{fillGapTracks} imputes missing values in all
\code{\link[trajectories]{Track}} objects of a
\code{\link[trajectories]{Tracks}} object. Gaps are filled
if their duration is $\le$ a user specified duration threshold
and if the distance between the spatial position of the last data
value before the gap and the spatial position of the first data
value after the gap is $\le$ a user specified distance threshold.
}
\examples{
#
}
\seealso{
\code{\link{reorganizeTracks}}, \code{\link{extractClutersBuffer}},
\code{\link{redefineIndices}},
\code{\link{fillGapTracks}}, \code{\link{locationsTrack}}.
}
