% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findStarts.R
\name{findStarts}
\alias{findStarts}
\title{Determines the start times of visits of a cluster.}
\usage{
findStarts(clusters, trs_spdf)
}
\arguments{
\item{clusters}{A \code{\link[sp]{SpatialGridDataFrame}} object containing
point clusters, as created with \code{\link{extractClusters}}.}

\item{trs_spdf}{A \code{\link[sp]{SpatialPointsDataFrame}} of the
household movement (as converted from a \code {\link[trajectories]{Tracks}}
object).}
}
\value{
A \code{POSIXct} vector of containing the starts (first points)
of the additional visits.
}
\description{
\code{findStarts} determines the start date/time of the latest visit
in a cluster/campsite if an additional visit has been detected by
\code{\link{searchNextVisit}}.
}
\details{
\code{\link{clusters}} should be an object of clusters/campsites as
returned by \code{\link{searchNextVisit}}. This function is used by
function \code{\link{clusterOrder}}.
}
\examples{
#
}
\seealso{
\code{\link{qTopology}}, \code{\link{pointsPerQuad}},
\code{\link{extractClusters}}, \code{\link{searchNextVisit}},
\code{\link{clusterOrder}}.
}
