% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateTrack.R
\name{gapdurationnextcampsite_wis}
\alias{gapdurationnextcampsite_wis}
\title{Computes Gap Durations Not Ignoring Short-Term Visits}
\usage{
gapdurationnextcampsite_wis(agg)
}
\arguments{
\item{agg}{A \code{data.frame} object as created internally in \code{\link{aggregateTrack}}.}
}
\value{
A numeric vector containing the duration of gaps between visits
as number of gap data values between visits, whereby short-term visits are not counted as
gaps and are considered as visits. Values for a following
move are always assigned to the last value of a
visit in \code{agg} as indicated by \code{agg$end}.
}
\description{
\code{gapdurationnextcampsite_wis} computes for a \code{data.frame} as created
internally in \code{\link{aggregateTrack}} the duration of gaps as number of
gap data values between visits, whereby short-term visits are not counted as
gaps and are considered as visits.
}
\examples{
#
}
\seealso{
\code{\link{aggregateTrack}}
}
\keyword{internal}
