% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.Track.R
\name{gapdurationnextcampsite_wos}
\alias{gapdurationnextcampsite_wos}
\title{Computes Gap Durations Ignoring Short-Term Visits}
\usage{
gapdurationnextcampsite_wos(agg)
}
\arguments{
\item{agg}{A \code{data.frame} object as created internally in \code{\link{aggregateTrack}}.}
}
\value{
A numeric vector containing the duration of gaps between campsites
as number of gap data values between campsite visits, whereby short-term
visits are counted as gaps. Values for a following
move are always assigned to the last value of a
campsite visit in \code{agg} as indicated by \code{agg$end & agg$campsite}.
}
\description{
\code{gapdurationnextcampsite_wos} computes for a \code{data.frame} as created
internally in \code{\link{aggregateTrack}} the duration of gaps as number of
gap data values between campsite visits, whereby short-term visits are counted as
gaps.
}
\examples{
#
}
\seealso{
\code{\link{aggregateTrack}}
}
\keyword{internal}
