% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapProportionTrack.R
\name{gapProportionTrack}
\alias{gapProportionTrack}
\title{Computes the proportion of gaps in a Track object.}
\usage{
gapProportionTrack(currenttrack)
}
\arguments{
\item{currenttrack}{An object of class \code{\link[trajectories:Track-class]{Track}}
with a variable \code{location}.}
}
\value{
A numeric value representing the proportion of gaps
in \code{currenttrack}.
}
\description{
\code{gapProportionTrack} computes the proportion of gaps in a
\code{\link[trajectories:Track-class]{Track}}.
}
\examples{
#
}
\seealso{
.
}
