% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNightConnectionsTrack.R
\name{getNightConnectionsTrack}
\alias{getNightConnectionsTrack}
\title{Extracts sub-daily connection values of GPS tracks.}
\usage{
getNightConnectionsTrack(currenttrack, tstart = 16, tend = 20)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories]{Track}} object.}

\item{tstart}{A numeric value indicating the start of the time interval
[h, values between 0 and 24]. The default is \code{tstart = 16}.}

\item{tend}{A numeric value indicating the end of the time interval
[h, values between 0 and 24]. The default is \code{tstart = 20}.}
}
\value{
The downsampled values of the slot \code{conections} of the
\code{\link[trajectories]{Track}} object (containing only data points
within the specified time interval).
}
\description{
\code{getNightConnectionsTrack} subsets the values of the slot
\code{connections} of a \code{\link[trajectories]{Track}} object by
extracting values that were
recorded within a certain time interval of a day (for example in order to
extract values that were recorded at night).
}
\examples{
#
}
\seealso{
\code{\link{downByDrop}}, \code{\link{downByDropTrs}},
\code{\link{getNightTrs}}, \code{\link{daynightFixesTrack}}.
}
