% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNightTrs.R
\name{getNightTrs}
\alias{getNightTrs}
\title{Extracts sub-daily values of GPS tracks.}
\usage{
getNightTrs(currenttrs, tstart = 16, tend = 20)
}
\arguments{
\item{tstart}{A numeric value indicating the start of the time interval
[h, values between 0 and 24]. The default is \code{tstart = 16}.}

\item{tend}{A numeric value indicating the end of the time interval
[h, values between 0 and 24]. The default is \code{tstart = 20}.}

\item{currentrs}{A \code{\link[trajectories:Track-class]{Tracks}} object.}
}
\value{
The downsampled \code{\link[trajectories:Track-class]{Tracks}} object
(containing only data points within the specified time interval).
}
\description{
\code{getNightTrs} subsets the values of all \code{\link[trajectories:Track-class]{Track}}
objects of a given \code{\link[trajectories:Track-class]{Tracks}}
object by extracting values that were recorded within a certain time interval
of a day (for example in order to extract values that were recorded at night).
}
\examples{
#
}
\seealso{
\code{\link{downByDrop}}, \code{\link{downByDropTrs}},
\code{\link{getNightTrack}}, \code{\link{daynightFixesTrack}},
\code{\link{daynightFixesTracks}},
\code{\link{getNightConnectionsTrack}}.
}
