% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herdersTA-package.R
\docType{package}
\name{herdersTA-package}
\alias{herdersTA}
\alias{herdersTA-package}
\title{herdersTA: Extracting Movement Characteristics from GPS Trajectories of Nomadic Households}
\description{
herdersTA provides functions in order to extract movement characteristics 
  from high-temporal resolution GPS trajectories of nomadic households. Key functions are 
  the extraction of visited locations using density-based clustering and flexible time 
  thresholding, the identification of individual visits, gap filling, temporal aggregation and the 
  extraction of raster and polygon data at locations. Moreover, functions to plot the 
  time course of visited locations with additional information while ensuring anonymisation 
  are provided.
}
\author{
\strong{Maintainer}: Henning Teickner \email{h_teic01@uni-muenster.de}

Authors:
\itemize{
  \item Christian Knoth \email{christian.knoth@uni-muenster.de}
}

}
\keyword{internal}
