% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyTimeIntervals.R
\name{identifyTimeIntervals}
\alias{identifyTimeIntervals}
\title{Assigns values of GPS tracks to time intervals.}
\usage{
identifyTimeIntervals(currenttrack, timeinterval = "month")
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories]{Track}} object with a
\code{POSIXct} column \code{time} in \code{currenttrack@data}. Data
values have to be regularly spaced (may be achieved for example with
\code{\link{reorganizeTracks}}).}

\item{timeinterval}{One of \code{"year"}, \code{"month"},
\code{"tendayinterval"}, \code{"day"} or \code{"all"}. If
\code{timeinterval = "year"}, data values will be classified on a
yearly resolution. If \code{timeinterval = "month"}, data values
will be classified on a monthly resolution. If
\code{timeinterval = "tendayinterval"}, data values will be classified
on a ten-day interval resolution (i.e. assigned to fixed ten-day intervals).
Each month has three ten-day intervals
with the last having varying lengths of 11 to 8 days (depending on the
month and the occurance of leap years). If \code{timeinterval = "day"},
data values will be classified on a daily resolution. All time intervals
that are not completely covered by \code{currenttrack} (i.e. potentially
at the start and end of \code{currenttrack}) are handled as if they would
represent the whole specified time interval. If \code{timeintervall = "all"},
all data values will be assigned to one class. Default is
\code{timeinterval = "month"}.}
}
\value{
An integer vector with a length equal to
\code{nrow(currenttrack@data)} with the same value for each data value
corresponding to the same time interval.
}
\description{
\code{identifyTimeIntervals} assign each data value (row) in
the \code{data} slot of a \code{\link[trajectories]{Track}}
object to a user specified time interval.
}
\examples{
#
}
\seealso{
\code{\link{removeDataTrack}}, \code{\link{removeDataTracks}},
\code{\link{nogapDurationTrack}}, \code{\link{nogapDurationTracks}}.
}
