% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latitudeDifferenceTrack.R
\name{latitudeDifferenceTrack}
\alias{latitudeDifferenceTrack}
\title{Computes the latitude difference between the most northern and most southern points of a Track object.}
\usage{
latitudeDifferenceTrack(currenttrack)
}
\arguments{
\item{currenttrack}{An object of class \code{\link[trajectories:Track]{Track}}
with a variable \code{altitude} in the data slot.}
}
\value{
a numeric value representing the latitudinal difference between the
most northern and most southern points in \code{currenttrack}.
}
\description{
\code{latitudeDifferenceTrack} computes the latitudinal difference between
the most northern and most southern points of a
\code{\link[trajectories:Track]{Track}}.
}
\examples{
#
}
\seealso{
.
}
