% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearityCampsiteLocationsTrack.R
\name{linearityCampsiteLocationsTrack}
\alias{linearityCampsiteLocationsTrack}
\title{Estimates the linearity of a Track object.}
\usage{
linearityCampsiteLocationsTrack(currenttrack)
}
\arguments{
\item{currenttrack}{An object of class \code{\link[trajectories:Track-class]{Track}}
with a variable \code{location} and \code{campsite}}
}
\value{
a numeric value representing the straightness index for
\code{currenttrack}.
}
\description{
\code{linearityCampsiteLocationsTrack} computes an estimate of the
linearity of a trajectory (object of class \code{\link[trajectories:Track-class]{Track}})
by (1) extracting the maximum distance between any two locations,
(2) extracting the maximum distance between any two remaining locations
orthogonal to the first line and (3) dividing the values extracted in (1) and (2).
The larger the value is the more linear is the trajectory overall.
}
\examples{
#
}
\seealso{
.
}
