% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTracks.R
\name{loadTracks}
\alias{loadTracks}
\title{Reads data from a set of GPS track files.}
\usage{
loadTracks(folder)
}
\arguments{
\item{folder}{A character value representing the path to a folder with .csv files
containing data on GPS tracks.}
}
\value{
A \code{\link[trajectories:Track-class]{Tracks}} object holding the data of the
GPS tracks.
}
\description{
\code{loadTracks} reads data from all .csv files containing data on
GPS tracks within a specified directory. The function also removes
empty .csv files within the directory by calling the function
\code{\link{removeEmptyFiles}}.
}
\examples{
#
}
\seealso{
\code{\link{removeEmptyFiles}}, \code{\link{readTrack}}.
}
