% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationsTrack.R
\name{locationsTrack}
\alias{locationsTrack}
\title{Identifies and classifies visits in GPS tracks.}
\usage{
locationsTrack(currenttrack, radius = 200, tmin = 345600,
  tmaxinterstices = 345600, timeinterval = 30 * 60, night = c(16,
  20),
  crs = "+proj=utm +zone=46 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0",
  summary = TRUE)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories]{Track}} object.}

\item{radius}{A numerical value representing the radius of the
buffers computed around each point [m] which are used for
clustering values to locations. Default is \code{radius = 800} [m].}

\item{tmin}{The minimum residence time at a specific location [s]
which is used to classify visits as campsites (if the residence
time at a specific location is larger than \code{tmin}) in
contrast to short-term visits of locations. The default is
\code{tmin = 345600}, i.e. 4 days.}

\item{tmaxinterstices}{The maximum time between subsequent visits
at the same location in case of which the duration of these visits
will be added in order to classify both visits together as long-term
visit (campsite) or short-term visit, based on \code{tmin}.}

\item{timeinterval}{A numerical value reperesenting the duration
of a time interval represented by one data value of
\code{currenttrack} [s].}

\item{night}{An integer vector with two elements:
\enumerate{
  \item The first element specifies the start hour of the night, e.g. \code{0}
  for 0 o'clock.
  \item The first element specifies the start hour of the night, e.g. \code{4}
  for 4 o'clock.
}}

\item{crs}{A character string describing a projection and datum
in the \code{PROJ.4} format (see \code{\link[rgdal]{projInfo}}).}

\item{summary}{Logical value indicating if the information on the
locations and visits should be summarised (\code{summary = TRUE})
or not (\code{summary = FALSE}). See the details section for further
information.}
}
\value{
\describe{
  \item{If (\code{summary = FALSE})}{A
  \code{\link[trajectories]{Track}} object that is identical
  to the input \code{\link[trajectories]{Track}} object, but
  has four additional columns in the \code{data} slot:
  \describe{
    \item {\code{location}}{An integer value for each identified
    spatial point cluster (location) increasing with the time starting
    from 1.}
    \item{\code{campsite}}{A logical value indicating if a visits of a
    location is classified as long-term visit (campsite) (\code{TRUE})
    or as short-term visit (\code{FALSE}).}
    \item{\code{visitsloc}}{An integer vector indicating the number of
    visits at a specific location the data point is assigned to (i.e.
    a counter for the visits at a specific location).}
    \item{\code{visitscampsite}}{An integer vector indicating the number
    of visits at a specific location the data point is assigned to that
    are classified as long-term visits (campsites) (i.e.
    a counter for the long-term visits at a specific location).}
  }
  Gaps, as indicated by the column \code{gap}, have \code{NA} values for
  all four variables.
  }
  \item{If (\code{summary = TRUE})}{A \code{data.frame}
  object summarising the locations and visits of the input
  \code{\link[trajectories]{Track}} object with the following variables:
  \describe{
    \item{location}{An integer value for each identified
    spatial point cluster (location) increasing with the time starting
    from 1.}
    \code{lon}{The longitude of the respective location (as mean value of
    the coordinates of the data values assigned to the visit).}
    \code{lat}{The latitude of the respective location (as mean value of
    the coordinates of the data values assigned to the visit).}
    \code{alt}{The altitude of the respective location (as mean value of
    the coordinates of the data values assigned to the visit).}
    \item{\code{visitsloc}}{An integer vector indicating the number of
    visits at a specific location the data point is assigned to (i.e.
    a counter for the visits at a specific location).}
    \item{campsite}{A logical value indicating if a visits of a
    location is classified as long-term visit (campsite) (\code{TRUE})
    or as short-term visit (\code{FALSE}).}
    \item{\code{visitscampsite}}{An integer vector indicating the number
    of visits at a specific location the data point is assigned to that
    are classified as long-term visits (campsites) (i.e.
    a counter for the long-term visits at a specific location).}
    \item{arrivaltime}{A \code{POSIXct} vector indicating the arrival time
    of the visit at the respective location.}
    \item{departuretime}{A \code{POSIXct} vector indicating the departure
    time of the visit from the respective location.}
    \item{residencetime}{A numerical vector indicating the residence time
    of each visit [s].}
    \code{speed}{The speed of the respective location (as mean value of
    the speed values of the data values assigned to the visit).}
  }
  }
}
}
\description{
\code{locationsTrack} identifies clusters (locations) of
points in GPS tracks (\code{\link[trajectories]{Track}}
object) based on their spatial proximity using
\code{\link{extractClustersBuffer}} and identifies individual
visits of the same cluster (location) along the track.
Additionally, visits are classified as long-term visits
(campsites) or short-term visits. During the identification of
clusters, data values within a user defined daily time interval
are considered. The remaining values are set to the next
location identified for the adjacent time intervals. This
procedure can be used in order to identify campsites by
assuming that a household has a campsite where it stayed
over night. If values outside the defined time interval
have a too far distance, they are classified as short-term
visit.
}
\details{
The function can be used in order to assign to each data
value of the input \code{\link[trajectories]{Track}}
object an id of the cluster it is assigned to (
\code{summary = FALSE}) or to summarise the information
for each visit of a location (\code{summary = TRUE}).
}
\examples{
#
}
\seealso{
\code{\link{reorganizeTracks}}, \code{\link{redefineIndices}},
\code{\link{fillGapTrack}}, \code{\link{fillGapTracks}},
\code{\link{extractClustersBuffer}}.
}
