% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationsTrack.R
\name{locationsTracks}
\alias{locationsTracks}
\title{Identifies and classifies visits in GPS tracks.}
\usage{
locationsTracks(currenttrack, radius = 200, tmin = 345600,
  summary = TRUE)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories]{Track}} object.}

\item{radius}{A numerical value representing the radius of the
buffers computed around each point [m] which are used for
clustering values to locations. Default is \code{radius = 800} [m].}

\item{tmin}{The minimum residence time at a specific location [s]
which is used to classify visits as campsites (if the residence
time at a specific location is larger than \code{tmin}) in
contrast to short-term visits of locations. The default is
\code{tmin = 345600}, i.e. 4 days.}

\item{summary}{Logical value indicating if the information on the
locations and visits should be summarised (\code{summary = TRUE})
or not (\code{summary = FALSE}). See the details section for further
information.}
}
\value{
A \code{\link[trajectories]{Track}} object that is:
\describe{
  \item{If (\code{summary = TRUE})}{identical to the input
  \code{\link[trajectories]{Track}} object, but has four additional
  columns in the \code{data} slot:
  \describe{
    \item {\code{location}}{An integer value for each identified
    spatial point cluster (location) increasing with the time starting
    from 1.}
    \item{\code{campsite}}{A logical value indicating if a visits of a
    location is classified as long-term visit (campsite) (\code{TRUE})
    or as short-term visit (\code{FALSE}).}
    \item{\code{visitsloc}}{An integer vector indicating the number of
    visits at a specific location the data point is assigned to (i.e.
    a counter for the visits at a specific location).}
    \item{\code{visitscampsite}}{An integer vector indicating the number
    of visits at a specific location the data point is assigned to that
    are classified as long-term visits (campsites) (i.e.
    a counter for the long-term visits at a specific location).}
  }
  Gaps, as indicated by the column \code{gap}, have \code{NA} values for
  all four variables.
  }
  \item{If (\code{summary = FALSE})}{A \code{\link[trajectories]{Track}}
  object summarising the locations and visits of the input
  \code{\link[trajectories]{Track}} object with the following variables:
  \describe{
    \item{location}{An integer value for each identified
    spatial point cluster (location) increasing with the time starting
    from 1.}
    \item{residencetime}{A numerical vector indicating the residence time
    of each visit [s].}
    \item{cmapsite}{A logical value indicating if a visits of a
    location is classified as long-term visit (campsite) (\code{TRUE})
    or as short-term visit (\code{FALSE}).}
    \item{arrivaltime}{A \code{POSIXct} vector indicating the arrival time
    of the visit at the respective location.}
    \item{departuretime}{A \code{POSIXct} vector indicating the departure
    time of the visit from the respective location.}
    \item{arrivaltimeindex}{An integer vector indicating the index of the
    data value of the input \code{\link[trajectories]{Track}} object that
    represents the arrival at the location for the respective visit.}
    \item{departuretimeindex}{An integer vector indicating the index of the
    data value of the input \code{\link[trajectories]{Track}} object that
    represents the departure from the location for the respective visit.}
    \item{\code{visitsloc}}{An integer vector indicating the number of
    visits at a specific location the data point is assigned to (i.e.
    a counter for the visits at a specific location).}
    \item{\code{visitscampsite}}{An integer vector indicating the number
    of visits at a specific location the data point is assigned to that
    are classified as long-term visits (campsites) (i.e.
    a counter for the long-term visits at a specific location).}
    \code{lon}{The longitude of the respective location (as mean value of
    the coordinates of the data values assigned to the visit).}
    \code{lat}{The latitude of the respective location (as mean value of
    the coordinates of the data values assigned to the visit).}
    \code{alt}{The altitude of the respective location (as mean value of
    the coordinates of the data values assigned to the visit).}
    \code{speed}{The speed of the respective location (as mean value of
    the speed values of the data values assigned to the visit).}
  }
  }
}
}
\description{
\code{locationsTrack} identifies clusters (locatons) of
points in GPS tracks (\code{\link[trajectories]{Track}}
object) based on their spatial proximity using
\code{\link{extractClustersBuffer}} and identifies individual
visits of the same cluster (location) along the track.
Additionally, visits are classified as long-term visits
(campsites) or short-term visits.
}
\details{
The function can be used in order to assign to each data
value of the input (\code{\link[trajectories]{Track}}
object an id of the cluster it is assigned to (
\code{summary = FALSE}) or to summarise the information
for each visit of a location (\code{summary = TRUE})
}
\examples{
#
}
\seealso{
\code{\link{reorganizeTracks}}, \code{\link{redefineIndices}},
\code{\link{fillGapTrack}}, \code{\link{fillGapTracks}},
\code{\link{extractClustersBuffer}}.
}
