% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudeDifferenceTrack.R
\name{longitudeDifferenceTrack}
\alias{longitudeDifferenceTrack}
\title{Computes the latitude difference between the most western and most eastern points of a Track object.}
\usage{
longitudeDifferenceTrack(currenttrack)
}
\arguments{
\item{currenttrack}{An object of class \code{\link[trajectories:Track-class]{Track}}
with a variable \code{altitude} in the data slot.}
}
\value{
a numeric value representing the longitudinal difference between the
most western and most eastern points in \code{currenttrack}.
}
\description{
\code{longitudeDifferenceTrack} computes the longitudinal difference between
the most western and most eastern points of a
\code{\link[trajectories:Track-class]{Track}}.
}
\examples{
#
}
\seealso{
.
}
