% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTracks.R
\name{mergeTracks}
\alias{mergeTracks}
\title{Merges \code{\link[trajectories]{Track}} objects of a \code{\link[trajectories]{Tracks}} object.}
\usage{
mergeTracks(currenttracks)
}
\arguments{
\item{currenttracks}{A \code{\link[trajectories]{Tracks}} object.}
}
\value{
A  \code{\link[trajectories]{Track}} object containing
the merged  \code{\link[trajectories]{Track}} objects of the
input  \code{\link[trajectories]{Tracks}} object.
}
\description{
\code{mergeTracks} merges the \code{\link[trajectories]{Track}}
objects of a \code{\link[trajectories]{Tracks}} objects into a
singular \code{\link[trajectories]{Track}} object.
}
\examples{
#
}
\seealso{
\code{\link{mergeTracksCollection}},
\code{\link{reorganizeTracks}}.
}
