% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTracksCollection.R
\name{mergeTracksCollection}
\alias{mergeTracksCollection}
\title{Merges \code{\link[trajectories:Track-class]{Track}} objects of a \code{\link[trajectories:Track-class]{Tracks}} objects.}
\usage{
mergeTracksCollection(trackscollection, cores = 1, clcall = NULL)
}
\arguments{
\item{cores}{An integer value representing the number of cores to
use in parallel computing.}

\item{clcall}{A function that is passed to
\code{\link[parallel]{clusterCall}}.}

\item{currenttrackscollection}{A
\code{\link[trajectories:Track-class]{TracksCollection}} object.}
}
\value{
A \code{\link[trajectories:Track-class]{Tracks}} object with
a \code{\link[trajectories:Track-class]{Track}} object for each former
\code{\link[trajectories:Track-class]{Tracks}} object containing
the merged  \code{\link[trajectories:Track-class]{Track}} objects of the
input  \code{\link[trajectories:Track-class]{Tracks}} object.
}
\description{
\code{mergeTracksCollection} merges the \code{\link[trajectories:Track-class]{Track}}
objects for each \code{\link[trajectories:Track-class]{Tracks}} object of a
\code{\link[trajectories:Track-class]{TracksCollection}} into a
singular \code{\link[trajectories:Track-class]{Track}} object.
}
\examples{
#
}
\seealso{
\code{\link{mergeTracks}}, \code{\link{reorganizeTracks}}.
}
