% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_dist_to_feature.R
\name{min_dist_to_feature}
\alias{min_dist_to_feature}
\title{Computes the minimum distance of locations to the nearest line or polygon border}
\usage{
min_dist_to_feature(currenttrack, feature, minrange = 10000)
}
\arguments{
\item{currenttrack}{A \code{Tracs} object usually representing movement of one household.
The \code{data} slot of \code{currenttrack} must contain a variable \code{"location"}
representing unique IDs for locations. Values with the location \code{0} receive
distance values of \code{NA}.}

\item{feature}{A SpatialLines or SpatialPolygons object containing the features
to compute distance to.}

\item{minrange}{A numerical value > 0 representing the minimum buffer radius within which
\code{min_dist_to_feature} initially searches for the closest polyline or polygon border
around a location. The unit of \code{minrange} is the same as for the reference system of
\code{currenttrack}.}
}
\value{
A \code{vector} with the shortest distance of all locations in \code{currenttrack}
to the closest feature in \code{feature}. For each value in \code{currenttrack}, a distance
value will be returned in the order of data values.
}
\description{
\code{min_dist_to_feature} computes the shortest distance of each location
(as identified by \code{\link{locationsTracks}}) in a \code{Tracks} object
to the closest polyline or polygon border. It does not include all possible
features, but starts looking for features in a certain distance from each
location. If no features are found, the search distance is increased
iteratively.
}
\examples{
#
}
