% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nogapDurationTrack.R
\name{nogapDurationTrack}
\alias{nogapDurationTrack}
\title{Determines the duration of gaps in GPS tracks.}
\usage{
nogapDurationTrack(currenttrack, timeinterval = 30 * 60)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories]{Track}} object with a
\code{boolean} column \code{gap} in \code{currenttrack@data} and an
\code{integer} column \code{timeinterval_id} in \code{currenttrack@data}. Data
values have to be regularly spaced (may be achieved for example with
\code{\link{reorganizeTracks}}).}

\item{timeinterval}{A numerical value reperesenting the duration of a
time interval represented by one data value of \code{currenttrack} [s].}
}
\value{
A \code{\link[trajectories]{Track}} object identical to
\code{currenttrack}, except for a new column
\code{currenttrack@data$duration_nogap}
representing the temporal duration of not missing values within the
corresponding time interval specified by
\code{currenttrack@data$id_timeinterval}.
}
\description{
\code{nogapDurationTrack} compute the temporal duration of not missing
values for time intervals of a \code{\link[trajectories]{Track}} object
as returned by \code{removeDataTrack}, based on the classification of
missing values as gaps by \code{reorganizeTracks} and on the
classification of time intervals to consider during analyses by
\code{removeDataTrack}.
}
\examples{
#
}
\seealso{
\code{\link{identifyTimeIntervals}}, \code{\link{removeDataTrack}},
\code{\link{removeDataTracks}}, \code{\link{nogapDurationTracks}}.
}
