% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.trackvisits.R
\name{plot.trackvisits}
\alias{plot.trackvisits}
\title{Plots Visits from a Track with Locations}
\usage{
plot.trackvisits(x, what = "lonlatanonymuous", seasons = data.frame(start =
  c(3, 5, 9, 11), end = c(4, 8, 10, 2), colour = c("yellow", "red",
  "burlywood1", "lightgray")), timethreshold = 0)
}
\arguments{
\item{x}{An object of class \code{link{trackvisits}} with the variable
\code{campsite} being defined.}

\item{what}{A character value defining which plot to create. See the description section.}

\item{seasons}{A \code{data.frame} object with a row for each season
and three columns:
\describe{
  \item{\code{start}}{A numeric value indicating the start month of the respective
  season.}
  \item{\code{start}}{A numeric value indicating the end month of the respective
  season.}
  \item{\code{colour}}{A character value indicating the colour used in order to fill
  the label boxes of visits with arrivals within the respective seasons.}
}}

\item{timethreshold}{A numeric value representing a time threshold of duarions between
adjacent visits [s]. If the duration between a visit and a next visit is larger than
\code{timethreshold}, this will be shown as dashed line in the plot.}
}
\value{
A  \code{\link[ggplot2]{ggplotobj}} object
}
\description{
\code{plot.trackvisits} is the plot function for objects of class
\code{\link{trackvisits}}. Several plots are possible:
\describe{
  \item{\code{"lonlat"}}{Plots the longitude on the x axis and the latitude
  on the y axis. Vists are plotted as points and are linked by lines. Points
  of visits at campsites are filled black, points of visits at short-term visits
  are filled white. Lines linking visits are plotted solid if the duration between
  visits does not exceed certain threshold values. Otherwise, they are plotted
  dashed. Each visit is indicated by a label that contains the following information:
  (1) step: The id of the visit, (2) loc.: The id of the location, (3) visit: The number
  of repeated campsite visits at the same location, (4) arr.: The start time of the
  visit (irrespective whether this is sure or not), (5) dep.: The end time of the
  visit (irrespective whether this is sure or not). Label boxes are filled according
  to an argument \code{"seasons}.}
  \item{\code{"lonlatanonymuous"}}{Like \code{"lonlat"}, but no labels will be
  added to the axes.}
  \item{\code{"altitude"}}{Plots the start time and end time of each visit on
  the x axis and the altitude of each visit on the y axis. Each visit is plotted
  as a pair of points linked by a solid straight line. Points
  of visits at campsites are filled black, points of visits at short-term visits
  are filled white. Lines linking visits are plotted solid if the duration between
  visits does not exceed certain threshold values. Otherwise, they are plotted
  dashed.Each visit is indicated by a label that contains the following information:
  (1) step: The id of the visit, (2) loc.: The id of the location, (3) visit: The number
  of repeated campsite visits at the same location, (4) arr.: The start time of the
  visit (irrespective whether this is sure or not), (5) dep.: The end time of the
  visit (irrespective whether this is sure or not). Label boxes are filled according
  to an argument \code{"seasons}.}
}
}
\examples{
#
}
\seealso{
\code{\link{reorganizeTracks}}, \code{\link{redefineIndices}},
\code{\link{fillGapTrack}}, \code{\link{fillGapTracks}},
\code{\link{extractClustersBuffer}}.
}
