% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLocationsTrackSummaryAltitude.R
\name{plotLocationsTrackSummaryAltitude}
\alias{plotLocationsTrackSummaryAltitude}
\title{Plots Altitude Profiles of Summarised Tracks as Returned by \code{\link{locaitonsTrack}}.}
\usage{
plotLocationsTrackSummaryAltitude(x, seasons = data.frame(start = c(3, 5, 9,
  11), end = c(4, 8, 10, 2), colour = c("yellow", "red", "burlywood1",
  "lightgray")))
}
\arguments{
\item{x}{An object as returned by \code{\link{locationsTrack}} with
parameter \code{summary = TRUE}.}

\item{seasons}{A \code{data.frame} object with a row for each season
and three columns:
\describe{
  \item{\code{start}}{A numeric value indicating the start month of the respective
  season.}
  \item{\code{start}}{A numeric value indicating the end month of the respective
  season.}
  \item{\code{colour}}{A character value indicating the colour used in order to fill
  the label boxes of visits with arrivals within the respective seasons.}
}}
}
\value{
A  \code{\link[ggplot2]{ggplotobj}} object.
}
\description{
\code{plotLocationsTrackSummaryAltitude} plots the result of \code{\link{locaitonsTrack}}
(with parameter  \code{summary = TRUE}) for a \code{\link[trajectories]{Track}}
object. This means that idividual campsites are plotted as dots linked with a
line. Labels indicate the step (i.e. the sequence of visits), the location id,
the arrival time and the altitude of the location. The fill of the label box indicates
the season of the arrival time. In contratst to \code{\link{plotLocationsTrackSummary}},
this function plots the altitude profile over the time for a \code{\link[trajectories]{Track}}
object.
}
\examples{
#
}
\seealso{
\code{\link{locationsTrack}}, \code{\link{locationsTracks}},
\code{\link{plotLocationsTrackSummary}}.
}
