% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointsPerQuad.R
\name{pointsPerQuad}
\alias{pointsPerQuad}
\title{Counts the number of points in the cells of a spatial grid.}
\usage{
pointsPerQuad(currenttracks, targetQsize)
}
\arguments{
\item{currenttracks}{A \code{\link[trajectories]{Tracks}} object.}

\item{targetQsize}{Numerical value representing the target size of the
quadrats to create.}
}
\value{
A \code{\link[sp]{SpatialGridDataFrame}} with quadrat cells
and the number of points within them.
}
\description{
\code{pointsPerQuad} determines the number of points per
cell (of a given size) in a \code{\link[trajectories]{Tracks}}
object.
}
\details{
This function is used by function \code{\link{extractClusters}} to
count the number of points within cells before determining clusters
of points.
}
\examples{
#
}
\seealso{
\code{\link{qTopology}}, \code{\link{extractClusters}},
\code{\link{searchNextVisit}}, \code{\link{findStarts}},
\code{\link{clusterOrder}}.
}
