% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qTopology.R
\name{qTopology}
\alias{qTopology}
\title{Defines the topology of a spatial grid.}
\usage{
qTopology(ppp, targetQsize)
}
\arguments{
\item{ppp}{A \code{\link[spatstat]{ppp}} (point pattern dataset) object.}

\item{targetQsize}{Numerical value representing the target size of the
quadrats to create.}
}
\value{
A list containing the size and number of cells in x and y.
}
\description{
\code{qTopology} determines topology for grid and quadratcount, i.e.
size in x and y such that the window of the ppp can be split in an
even number of quadrats, approximating a certain target edge length,
e.g. 1000 m.
}
\details{
This function is used by function \code{\link{pointsPerQuad}}
to create the cells in which to count the number of points.
}
\examples{
#
}
\seealso{
\code{\link{pointsPerQuad}}, \code{\link{extractClusters}},
\code{\link{searchNextVisit}}, \code{\link{findStarts}},
\code{\link{clusterOrder}}.
}
