% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTrack.R
\name{readTrack}
\alias{readTrack}
\title{Reads data from an individual GPS track file.}
\usage{
readTrack(fname)
}
\arguments{
\item{fname}{A character value representing the file name of the .csv file
containing the GPS data to read.}
}
\value{
A \code{\link[trajectories]{Track}} object holding the data of the
GPS track.
}
\description{
\code{readTrack} reads data from a .csv file containing data on an
individual GPS track.
}
\examples{
#
}
\seealso{
\code{\link{removeEmptyFiles}}.
}
