% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redefineIndices.R
\name{redefineIndices}
\alias{redefineIndices}
\title{Redefines indices based on a time vector.}
\usage{
redefineIndices(df, indices, time, notchange = NULL)
}
\arguments{
\item{df}{A \code{data.frame} object with a column representing
indices and a column representing time information as \code{POSIXct}
values.}

\item{indices}{A character value giving the name of the column of
\code{df} which contains the time values.}

\item{time}{indices A character value giving the name of the column of
\code{df} which contains the index values.}

\item{notchange}{An integer vector specifying which original index values
should not be changed. Default is \code{notchange = NULL}, i.e. all indices
will be changed.}
}
\value{
A \code{data.frame} object with the redefined index values.
}
\description{
\code{redefineIndices} redefines a column with index values (integer)
according to a column with time information in a  \code{data.frame}
object, i.e. new indices will be assigned to each set of rows with
the same index in relation to their order in time.
}
\examples{
#
}
\seealso{
\code{\link{reorganizeTracks}}, \code{\link{extractClutersBuffer}},
\code{\link{fillGapTrack}}, \code{\link{fillGapTracks}},
\code{\link{locationsTrack}}.
}
