% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeEmptyFolders.R
\name{removeEmptyFolders}
\alias{removeEmptyFolders}
\title{Finds non-empty folders in a directory.}
\usage{
removeEmptyFolders(dirlist)
}
\arguments{
\item{dirlist}{A list of folder names that should be chekced.}
}
\value{
A list of folder names that are not empty.
}
\description{
\code{removeEmptyFolders} returns the names of non-empty folders within a
directory by ignoring empty folders.
}
\examples{
#
}
\seealso{
\code{\link{removeEmptyFiles}}.
}
