% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganizeTracks.R
\name{reorganizeTracks}
\alias{reorganizeTracks}
\title{Reorganises \code{\link[trajectories:Track-class]{Tracks}} objects.}
\usage{
reorganizeTracks(currenttracks, interval = 60 * 30, tz = "GMT",
  crs = "+proj=longlat +ellps=WGS84", cores = 1, clcall = NULL)
}
\arguments{
\item{currenttracks}{A \code{\link[trajectories:Track-class]{Tracks}} object.}

\item{interval}{A numerical value representing the length of the
time interval to assign all data values to [s]. Default is
\code{interval = 60*30}, i.e. 30 minutes.}

\item{tz}{A character value indicating the time zone the temporal
information orresponds to. Default is \code{tz = "GMT"}.}

\item{crs}{A character string describing a projection and datum
in the \code{PROJ.4} format (see \code{\link[rgdal]{projInfo}}).}

\item{cores}{An integer value representing the number of cores to
use in parallel computing.}

\item{clcall}{A function that is passed to
\code{\link[parallel]{clusterCall}}.}
}
\value{
A \code{\link[trajectories:Track-class]{Tracks}} object containing
the reorganised \code{\link[trajectories:Track-class]{Track}} objects of the
input A \code{\link[trajectories:Track-class]{Tracks}} object.
}
\description{
\code{reorganizeTracks} reorganises a \code{\link[trajectories:Track-class]{Tracks}}
object with several \code{\link[trajectories:Track-class]{Track}} objects
covering roughly the same time period by assigning data points to
the same overall time interval and including gaps as new data
points within each \code{\link[trajectories:Track-class]{Track}} object.
}
\details{
If there are more than one data point per interval and
\code{\link[trajectories]{Track-class}} object, only the first data
point within this interval is retained.
}
\examples{
#
}
\seealso{
\code{\link{mergeTracksCollection}}, \code{\link{mergeTracks}}.
}
