% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchNextVisit.R
\name{searchNextVisit}
\alias{searchNextVisit}
\title{Determines the number of visits of a cluster.}
\usage{
searchNextVisit(clusters, trs_spdf)
}
\arguments{
\item{clusters}{A \code{\link[sp:SpatialPixelsDataFrame-class]{SpatialGridDataFrame}} object containing
point clusters, as created with \code{\link{extractClusters}}.}

\item{trs_spdf}{A \code{\link[sp:SpatialPoints]{SpatialPointsDataFrame}} of the
household movement (as converted from a \code{\link[trajectories:Track-class]{Tracks}}
object).}
}
\value{
A \code{\link[sp:SpatialPoints]{SpatialPointsDataFrame}} holding the
clusters/campsites with updated \code{times_visited}.
}
\description{
\code{searchNextVisit} checks whether the clusters/campsites have been
visited more than the currently known visits (\code{times_visited}).
Updates \code{times_visited} and determines the end (latest point in cell)
of the current visit. This function is used by function `
\code{\link{clusterOrder}}.
}
\examples{
#
}
\seealso{
\code{\link{qTopology}}, \code{\link{pointsPerQuad}},
\code{\link{extractClusters}}, \code{\link{findStarts}},
\code{\link{clusterOrder}}.
}
