% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siTrack.R
\name{siTrack}
\alias{siTrack}
\title{Computes various summary indicators for a Track object.}
\usage{
siTrack(currenttrack, fun = stats::median)
}
\arguments{
\item{currenttrack}{An object of class \code{\link[trajectories:Track-class]{Track}}
with a variable \code{location} and \code{campsite}}

\item{fun}{One of \code{mean} or \code{median)}, depending on which function should be
used in order to compute average values.}

\item{campsite}{A logical value indicating if only campsites should be considered
(\code{TRUE}) or any locations (\code{FALSE}).}
}
\value{
a numeric value representing the number of unique (campsite) locations
in \code{currenttrack}.
}
\description{
\code{siTrack} is a wrapper function for functions that takes a \code{Track}
object as input and return numeric values/vectors as output that summarise the
\code{Track} object. The functions are:
\enumerate{
  \item \code{\link{averageLocationsAltitudeTrack}}
  \item \code{\link{altitudeRangeTrack}}
  \item \code{\link{altitudeDifferenceTrack}}
  \item \code{\link{averageDistanceCampsiteLocationsTrack}}
  \item \code{\link{longitudeDifferenceTrack}}
  \item \code{\link{latitudeDifferenceTrack}}
  \item \code{\link{sumLongitudeTrack}}
  \item \code{\link{sumLatitudeTrack}}
  \item \code{\link{chullTrack}}: the area and circumference are extracted
  from the \code{SpatialPolygons} object.
  \item \code{\link{averageDirectionCampsiteLocationsTrack}}
  \item \code{\link{countUniqueLocationsTrack}} (number of unique campsite locations
  and number of campsite locations with repeated campsite visits).
  \item \code{\link{gapProportionTrack}}
}
}
\examples{
#
}
\seealso{
.
}
