% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straighntnessindexCampsiteLocationsTrack.R
\name{straighntnessindexCampsiteLocationsTrack}
\alias{straighntnessindexCampsiteLocationsTrack}
\title{Computes the straightness index for the trajectory between the first and last campsite location in a Track object.}
\usage{
straighntnessindexCampsiteLocationsTrack(currenttrack)
}
\arguments{
\item{currenttrack}{An object of class \code{\link[trajectories:Track-class]{Track}}
with a variable \code{location} and \code{campsite}}
}
\value{
a numeric value representing the straightness index for
\code{currenttrack}.
}
\description{
\code{straighntnessindexCampsiteLocationsTrack} computes the straightness index
for the trajectory between the first and last locations of a
\code{\link[trajectories:Track-class]{Track}} in the order of movement.
}
\examples{
#
}
\seealso{
.
}
