% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumLatitudeTrack.R
\name{sumLatitudeTrack}
\alias{sumLatitudeTrack}
\title{Computes the sum of the latitudinal transitions of a Track object.}
\usage{
sumLatitudeTrack(currenttrack)
}
\arguments{
\item{currenttrack}{An object of class \code{\link[trajectories:Track-class]{Track}}
with a variable \code{location} and \code{campsite} in the data slot.}
}
\value{
a numeric value representing the sum of the latitude transitions
for \code{currenttrack}.
}
\description{
\code{sumLatitudeTrack} sums the latitude differences between adjacent campsite
locations for a \code{\link[trajectories:Track-class]{Track}}.
}
\examples{
#
}
\seealso{
.
}
