% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumLongitudeTrack.R
\name{sumLongitudeTrack}
\alias{sumLongitudeTrack}
\title{Computes the sum of the longitudinal transitions of a Track object.}
\usage{
sumLongitudeTrack(currenttrack)
}
\arguments{
\item{currenttrack}{An object of class \code{\link[trajectories:Track-class]{Track}}
with a variable \code{location} and \code{campsite} in the data slot.}
}
\value{
a numeric value representing the sum of the longitude transitions
for \code{currenttrack}.
}
\description{
\code{sumLongitudeTrack} sums the longitude differences between adjacent campsite
locations for a \code{\link[trajectories:Track-class]{Track}}.
}
\examples{
#
}
\seealso{
.
}
