% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryIndicatorsCampsitesTrack_arrivaltime.R
\name{summaryIndicatorsCampsitesTrack_arrivaltime}
\alias{summaryIndicatorsCampsitesTrack_arrivaltime}
\title{Extracts a Vector of Matched Unique Values for Campsites.}
\usage{
summaryIndicatorsCampsitesTrack_arrivaltime(currenttrack)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories:Track-class]{Track}} object that
has been processed as described in the description. Additionally, the function
needs a logical variable indicating if the household arrived at a specific day
(\code{TRUE}) or not (\code{FALSE}) as \code{arrived}.}
}
\value{
A \code{data.frame} object with a row for each campsite and two
columns:
\describe{
  \item{campsite}{An identifier for each unique campsite as provided by
  \code{currenttrack$aggregatedcampsitevists.}}
  \item{arrival}{A character vector representing the arrival days for each
  campsite.}
}
}
\description{
\code{summaryIndicatorsCampsitesTrack_arrivaltime} extracts for each
campsite the day when the household arrived. If this time point is
not sure, the function returns \code{NA}. The function can  be used in
combination with a
\code{\link[trajectories:Track-class]{Track}} object as returned by
\code{\link{aggregateDailyLocationsTrack}} for which unique campsite
visits are identified by a variable \code{aggregatedcmapsitevisits}.
}
\examples{
#
}
\seealso{
\code{\link{removeDataTracks}}, \code{\link{nogapDurationTracks}}.
}
