% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryIndicatorsCampsitesTrack_departuretime.R
\name{summaryIndicatorsCampsitesTrack_departuretime}
\alias{summaryIndicatorsCampsitesTrack_departuretime}
\title{Extracts a Vector of Matched Unique Values for Fixed Ten-Day Intervals}
\usage{
summaryIndicatorsCampsitesTrack_departuretime(currenttrack)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories:Track-class]{Track}} object that
has been processed as described in the description. Additionally, the function
needs a logical variable indicating if the household left at a specific day
(\code{TRUE}) or not (\code{FALSE}) as \code{left}.}
}
\value{
A vector \code{data.frame} object with a row for each campsite and two
columns:
\describe{
  \item{campsite}{An identifier for each unique campsite as provided by
  \code{currenttrack$aggregatedcampsitevists.}}
  \item{departure}{A character vector representing the departure days for each
  campsite.}
}
}
\description{
\code{summaryIndicatorsCampsitesTrack_departuretime} extracts for each
campsite the day when the household left. If this time point is
not sure, the function returns \code{NA}. The function can  be used in
combination with a
\code{\link[trajectories:Track-class]{Track}} object as returned by
\code{\link{aggregateDailyLocationsTrack}} for which unique campsite
visits are identified by a variable \code{aggregatedcmapsitevisits}.
}
\examples{
#
}
\seealso{
\code{\link{removeDataTracks}}, \code{\link{nogapDurationTracks}}.
}
