% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryIndicatorsCampsitesTrack_distance.R
\name{summaryIndicatorsCampsitesTrack_distance}
\alias{summaryIndicatorsCampsitesTrack_distance}
\title{Extracts Linear Distances Between Campsite Locations.}
\usage{
summaryIndicatorsCampsitesTrack_distance(currenttrack)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories:Track-class]{Track}} object that
has been processed as described in the description.}
}
\value{
A \code{data.frame} object with a row for each campsite and two
columns:
\describe{
  \item{campsite}{An identifier for each unique campsite as provided by
  \code{currenttrack$aggregatedcampsitevists.}}
  \item{distance}{A numeric vector representing the linear distance
  between the current campsite (indicated by \code{campsite}) and the
  following campsite.}
}
}
\description{
\code{summaryIndicatorsCampsitesTrack_distance} extracts for each
campsite the linear distance to the following campsite. For the first
campsite the value is set to \code{NA} in order to indicate that the
distance to the previous campsite is unknown. The
function can  be used in combination with a
\code{\link[trajectories:Track-class]{Track}} object as returned by
\code{\link{aggregateDailyLocationsTrack}} for which unique campsite
visits are identified by a variable \code{aggregatedcmapsitevisits}.
}
\examples{
#
}
\seealso{
\code{\link{removeDataTracks}}, \code{\link{nogapDurationTracks}}.
}
