% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryIndicatorsCampsitesTrack_endtime.R
\name{summaryIndicatorsCampsitesTrack_endtime}
\alias{summaryIndicatorsCampsitesTrack_endtime}
\title{Extracts a Vector of Matched Unique Values for Campsites.}
\usage{
summaryIndicatorsCampsitesTrack_endtime(currenttrack)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories:Track-class]{Track}} object that
has been processed as described in the description.}
}
\value{
A \code{data.frame} object with a row for each campsite and two
columns:
\describe{
  \item{campsite}{An identifier for each unique campsite as provided by
  \code{currenttrack$aggregatedcampsitevists.}}
  \item{end}{A character vector representing the days for which the
  last data values are available for each campsite.}
}
}
\description{
\code{summaryIndicatorsCampsitesTrack_endtime} extracts for each
campsite the day for which the last data values are available. The
function can  be used in combination with a
\code{\link[trajectories:Track-class]{Track}} object as returned by
\code{\link{aggregateDailyLocationsTrack}} for which unique campsite
visits are identified by a variable \code{aggregatedcmapsitevisits}.
}
\examples{
#
}
\seealso{
\code{\link{removeDataTracks}}, \code{\link{nogapDurationTracks}}.
}
