% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryIndicatorsCampsitesTrackGeneric.R
\name{summaryIndicatorsCampsitesTrackGeneric}
\alias{summaryIndicatorsCampsitesTrackGeneric}
\title{Computes summary indicators for GPS tracks.}
\usage{
summaryIndicatorsCampsitesTrackGeneric(
  currenttrack,
  normalise = TRUE,
  fun = mean,
  what
)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories:Track-class]{Track}} object that
has been processed as described in the description.}

\item{normalise}{A logical value indicating if the computed summary
indicators should be normalized relative to the duration of data values
(as specified by the variable \code{data$nogapduration} of \code{currenttrack})
for each fixed
ten-day interval (\code{normalize = TRUE}) or not (\code{normalize = FALSE}).}

\item{fun}{A function that takes a numeric vector as input and returns
a numeric value that is used in order to temporally aggregate the values.}

\item{what}{A numeric vector with indices of the variables in
\code{currenttrack@data} to summarise.}
}
\value{
A \code{data.frame} object with with each row representing a campsite
and two columns
\describe{
  \item{\code{campsite}}{A character vector indicating the campsite (combination
  of the location number and the number of repeated visits).}
  \item{\code{value}}{A vector with the aggregated values.}

}
}
\description{
\code{summaryIndicatorsCampsitesTrackGeneric} can be used to compute various
summary indicators for a \code{\link[trajectories:Track-class]{Track}} object
that has been processed with \code{\link{reorganizeTracks}},
\code{\link{locationsTrack}} and \code{\link{aggregateDailyLocationsTrack}}.
Summary indicators are computed for each campsite.
\code{summaryIndicatorsIntervalsTrackGeneric}
does not compute special summary indicators in a non-generic way, but
simply applies a user-defined function to a selection of variables of a
\code{\link[trajectories:Track-class]{Track}} object.
}
\examples{
#
}
\seealso{
\code{\link{removeDataTracks}}, \code{\link{nogapDurationTracks}}.
}
