% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryIndicatorsIntervalsTrack_altitude.R
\name{summaryIndicatorsIntervalsTrack_altitude}
\alias{summaryIndicatorsIntervalsTrack_altitude}
\title{Computes summary indicators for GPS tracks.}
\usage{
summaryIndicatorsIntervalsTrack_altitude(currenttrack, normalise = FALSE)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories:Track-class]{Track}} object that
has been processed as described in the description.}

\item{normalise}{A logical value indicating if the computed summary
indicators should be normalized relative to the duration of data values
(as specified by \code{currenttrack@data$nogapduration}) for each fixed
ten-day interval (\code{normalize = TRUE}) or not (\code{normalize = FALSE}).}
}
\value{
A numeric vector with the (normalised) altitudinal height
differences covered within each fixed ten-day interval \code{currenttrack}
covers.
}
\description{
\code{summaryIndicatorsIntervalsTrack_altitude} computes the sum of
the height differences (i.e. all covered height meters)
a household covered during a fixed ten-day interval. Height
distances are
computed for each value within a fixed ten-day interval and additionally
 between the first value of the following interval and the
last value of the target interval. The function requires a
\code{\link[trajectories:Track-class]{Track}} object as returned by
\code{\link{aggregateDailyLocationsTrack}}. Hence, only linear
altitudeinal differences between campsites are computed.
}
\examples{
#
}
\seealso{
\code{\link{removeDataTracks}}, \code{\link{nogapDurationTracks}}.
}
