% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryIndicatorsIntervalsTrack_categorylist.R
\name{summaryIndicatorsIntervalsTrack_categorylist}
\alias{summaryIndicatorsIntervalsTrack_categorylist}
\title{Extracts a Vector of Matched Unique Values for Fixed Ten-Day Intervals}
\usage{
summaryIndicatorsIntervalsTrack_categorylist(currenttrack, what,
  whatcondition = NULL, conditionvalue = TRUE)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories:Track-class]{Track}} object that
has been processed as described in the description.}

\item{what}{A numeric value indicating a variable in the \code{data} slot of
\code{currenttrack} for which to extract the categories.}

\item{whatcondition}{An optional numeric value indicating a variable in the
\code{data} slot of \code{currenttrack} that should function as condition in order
to subset the values of the variable represented by \code{what}.}

\item{conditionvalue}{A value of the variable indicated by \code{whatcondition}.
Will only be considered if \code{!is.null(whatcondition)}. The values of the variable
represented by \code{what} will be filtered for \code{conditionvalue} in the variable
indicated by \code{whatcondition} prior summarising.}
}
\value{
A vector of JSON representing the unique values of the selected categorial
variable occuring within each fixed ten-dy intervals.
}
\description{
\code{summaryIndicatorsIntervalsTrack_categorylist} extracts for each
fixed ten-day interval a vector of unique values of a categorial variable,
sorted according to the temporal order of these values within each
fixed ten-day interval. The function requires a
\code{\link[trajectories:Track-class]{Track}} object as returned by
\code{\link{aggregateDailyLocationsTrack}}.
}
\examples{
#
}
\seealso{
\code{\link{removeDataTracks}}, \code{\link{nogapDurationTracks}}.
}
