% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryIndicatorsIntervalsTrack_distance.R
\name{summaryIndicatorsIntervalsTrack_distance}
\alias{summaryIndicatorsIntervalsTrack_distance}
\title{Computes summary indicators for GPS tracks.}
\usage{
summaryIndicatorsIntervalsTrack_distance(currenttrack, normalise = FALSE)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories]{Track}} object that
has been processed as described in the description.}

\item{normalise}{A logical value indicating if the computed summary
indicators should be normalized relative to the duration of data values
(as specified by \code{currenttrack@data$nogapduration}) for each fixed
ten-day interval (\code{normalize = TRUE}) or not (\code{normalize = FALSE}).}
}
\value{
A numeric vector with the (normalised) distance covered within
each fixed ten-day interval \code{currenttrack} covers.
}
\description{
\code{summaryIndicatorsIntervalsTrack_distance} computes the distance
a household covered during a fixed ten-day interval. Distances are
computed for each value within a fixed ten-day interval and additionally
for the distance between the last value of the previous interval and the 
first value of the target interval. The function requires a 
\code{\link[trajectories]{Track}} object as returned by 
\code{\link{aggregateDailyLocationsTrack}}. Hence, only linear
distances between campsites are computed.
}
\examples{
#
}
\seealso{
\code{\link{removeDataTracks}}, \code{\link{nogapDurationTracks}}.
}
