% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryIndicatorsIntervalsTrack_timesmoved.R
\name{summaryIndicatorsIntervalsTrack_timesmoved}
\alias{summaryIndicatorsIntervalsTrack_timesmoved}
\title{Computes summary indicators for GPS tracks.}
\usage{
summaryIndicatorsIntervalsTrack_timesmoved(currenttrack, normalise = FALSE)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories:Track-class]{Track}} object that
has been processed as described in the description.}

\item{normalise}{A logical value indicating if the computed summary
indicators should be normalized relative to the duration of data values
(as specified by \code{currenttrack@data$nogapduration}) for each fixed
ten-day interval (\code{normalize = TRUE}) or not (\code{normalize = FALSE}).}
}
\value{
A numeric vector with the (normalised) total altitudinal height
differences covered within each fixed ten-day interval \code{currenttrack}
covers.
}
\description{
\code{summaryIndicatorsIntervalsTrack_timesmoved} computes the number
of campsite moves during a fixed ten-day interval. For each fixed
ten-day interval, the maximum number of arrivals OR departures OR
distinct campsite visits (i.e. at different locations) is considered
as the number of moves. This is done because exact arrival or departure
times can only be given if there is no gap (that is too long) before
a data value at a new location. The function requires a
\code{\link[trajectories:Track-class]{Track}} object as returned by
\code{\link{aggregateDailyLocationsTrack}}. Hence, only linear
altitudeinal differences between campsites are computed.
}
\examples{
#
}
\seealso{
\code{\link{removeDataTracks}}, \code{\link{nogapDurationTracks}}.
}
