% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryIndicatorsIntervalsTrackGeneric.R
\name{summaryIndicatorsIntervalsTrackGeneric}
\alias{summaryIndicatorsIntervalsTrackGeneric}
\title{Computes summary indicators for GPS tracks.}
\usage{
summaryIndicatorsIntervalsTrackGeneric(currenttrack, normalise = TRUE,
  fun = mean, what)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories]{Track}} object that
has been processed as described in the description.}

\item{normalise}{A logical value indicating if the computed summary
indicators should be normalized relative to the duration of data values
(as specified by \code{currenttrack@data$nogapduration}) for each fixed
ten-day interval (\code{normalize = TRUE}) or not (\code{normalize = FALSE}).}

\item{fun}{A function that takes a numeric vector as input and returns
a numeric value that is used in order to temporally aggregate the values.}

\item{what}{A numeric vector with indices of the variables in
\code{currenttrack@data} to summarise.}
}
\description{
\code{summaryIndicatorsIntervalsTrackGeneric} computes various summary indicators
for a \code{\link[trajectories]{Track}} object that has been processed
with \code{\link{reorganizeTrack}}, \code{\link{locationsTrack}} and
\code{\link{aggregateDailyLocationsTrack}}. Summary
indicators are computed for a temporal resolution of fixed ten-day intervals
and additional for the transition between the first vale before
a ten-day interval and the first value of the ten-day interval. In contrast to
\code{summaryIndicatorsIntervalsTrack}, \code{summaryIndicatorsIntervalsTrackGeneric}
does not compute special summary indicators in a non-generic way, but
simply applies a user-defined function to a selection of variables of a
\code{\link[trajectories]{Track}} object.
}
\examples{
#
}
\seealso{
\code{\link{removeDataTracks}}, \code{\link{nogapDurationTracks}}.
}
