% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeActiveTrack.R
\name{timeActiveTrack}
\alias{timeActiveTrack}
\title{Analyses the temporal coverage of GPS tracks.}
\usage{
timeActiveTrack(currenttrack, tstart, tend)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories:Track-class]{Track}} object.}

\item{tstart}{A \code{POSIXct} object indicating the date and time of the start
of the target period.}

\item{tend}{A \code{POSIXct} object indicating the date and time of the end of
the target period.}
}
\value{
A numerical value:
\itemize{
  \item 0 If the track does not cover the time period.
  \item 1 If the track touches the time period (i.e. the maximum time value of
  the track equals the start value of the time period or the minimum time
  value of the track equals the end value of the time period).
  \item The number of seconds the time period and the track intersect in any
  other case.
}
}
\description{
\code{timeActiveTrack} computes the time [s] during which position data was
collected in a given \code{\link[trajectories:Track-class]{Track}} object (for an
individual \code{\link[trajectories:Track-class]{Track}} object) within a given time period.
}
\examples{
#
}
\seealso{
\code{\link{timeActiveTracks}}, \code{\link{timeActiveTracksCollection}}.
}
