% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeActiveTracks.R
\name{timeActiveTracks}
\alias{timeActiveTracks}
\title{Analyses the temporal coverage of GPS tracks.}
\usage{
timeActiveTracks(currenttracks, tstart, tend)
}
\arguments{
\item{currenttracks}{A \code{\link[trajectories]{Tracks}} object.}

\item{tstart}{A \code{POSIXct} object indicating the date and time of the start
of the target period.}

\item{tend}{A \code{POSIXct} object indicating the date and time of the end of
the target period.}
}
\value{
A numerical value representing the summed values of
\code{timeActiveTrack} for all \code{\link[trajectories]{Track}} objects of the
specified \code{\link[trajectories]{Tracks}} object. This value corresponds to
the number of seconds the \code{\link[trajectories]{Tracks}} object and the time
period intersect.
}
\description{
\code{timeActiveTracks} computes the time [s] during which position data was
collected in all \code{\link[trajectories]{Track}} objects of a specified
\code{\link[trajectories]{Tracks}} object within a given time period. The
coverage of all \code{\link[trajectories]{Track}} objects is summed.
}
\examples{
#
}
\seealso{
\code{\link{timeActiveTrack}}, \code{\link{timeActiveTrackCollection}}.
}
