% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeActiveTracksCollection.R
\name{timeActiveTracksCollection}
\alias{timeActiveTracksCollection}
\title{Analyses the temporal coverage of GPS tracks.}
\usage{
timeActiveTracksCollection(trcollection, tstart, tend)
}
\arguments{
\item{trcollection}{A \code{\link[trajectories:Track-class]{TracksCollection}} object.}

\item{tstart}{A \code{POSIXct} object indicating the date and time of the start
of the target period.}

\item{tend}{A \code{POSIXct} object indicating the date and time of the end of
the target period.}
}
\value{
A numerical value representing the proportional coverage of the
intersection between each \code{\link[trajectories:Track-class]{Tracks}} object and the
specified time period relative to the total duration of the specified time
period.
}
\description{
\code{timeActiveTrackCollection} computes the time [s] during which position data
was collected in all \code{\link[trajectories:Track-class]{Tracks}} objects of a specified
\code{\link[trajectories:Track-class]{TracksCollection}} object within a given time period.
}
\examples{
#
}
\seealso{
\code{\link{timeActiveTrack}}, \code{\link{timeActiveTracks}}.
}
