% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalDistanceTrack.R
\name{totalDistanceTrack}
\alias{totalDistanceTrack}
\title{Computes the total distance covered by a Track object.}
\usage{
totalDistanceTrack(currenttrack)
}
\arguments{
\item{currenttrack}{An object of class \code{\link[trajectories:Track-class]{Track}}.}
}
\value{
a numeric value representing the total distance covered by \code{currenttrack}.
}
\description{
\code{totalDistanceTrack} computes the total distance covered by a
\code{\link[trajectories:Track-class]{Track}}. This is the sum of all distance
values in the \code{connections} slot.
}
\examples{
#
}
\seealso{
.
}
