% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalGapDuration.R
\name{totalGapDuration}
\alias{totalGapDuration}
\title{Computes the duration of gaps in GPS tracks.}
\usage{
totalGapDuration(
  currenttracks,
  tstart = min(currenttracks@tracksData$tmin),
  tend = max(currenttracks@tracksData$tmax),
  threshold = 18000,
  tolerance = 2
)
}
\arguments{
\item{currenttracks}{A \code{\link[trajectories:Track-class]{Tracks}} object.}

\item{tstart}{A \code{POSIXct} object indicating the date and time of the start
of the target period. Default is the first value of the
\code{\link[trajectories:Track-class]{Track}} object.}

\item{tend}{A \code{POSIXct} object indicating the date and time of the end of
the target period. Default is the last value of the
\code{\link[trajectories:Track-class]{Track}} object.}

\item{threshold}{An integer value indicating the threshold for the time between
two data points [s] above which this is considered as a gap in the
\code{\link[trajectories:Track-class]{Track}} object. Default is \code{threshold = 18000},
i.e. 5 hours. If \code{threshold = NULL}, the threshold value is calculated from
the data as \code{mean(duration) + tolerance * sd(duration)}.}

\item{tolerance}{An integer value indicating the tolerance of the threshold value
calculated from the data if \code{threshold = NULL}. Default is
\code{tolerance = 2}, i.e. 2 seconds.}
}
\value{
A numeric value representing the sum of the duration of gaps within the
\code{\link[trajectories:Track-class]{Tracks}} object [s].
}
\description{
\code{gapDuration} computes the duration of gaps [s] for all
\code{\link[trajectories:Track-class]{Track}} objects within a given
\code{\link[trajectories:Track-class]{Tracks}} object and within a given time
period, using the function \code{\link{gapDuration}}.
}
\examples{
#
}
\seealso{
\code{\link{findGaps}}, \code{\link{countGaps}},
\code{\link{gapProportion}}, \code{\link{totalGapProportion}}.
}
