% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalRelativeAltitudeDistanceTrack.R
\name{totalRelativeAltiduinalDistanceTrack}
\alias{totalRelativeAltiduinalDistanceTrack}
\title{Computes the total altitudinal distance covered by a Track object.}
\usage{
totalRelativeAltiduinalDistanceTrack(currenttrack)
}
\arguments{
\item{currenttrack}{An object of class \code{\link[trajectories:Track]{Track}}
with a variable \code{altitude} in the \code{data} slot.}
}
\value{
a numeric value representing the total altitudinal distance covered by
\code{currenttrack}.
}
\description{
\code{totalRelativeAltitudeDistanceTrack} computes the total altitudinal distance
covered by a \code{\link[trajectories:Track]{Track}} (sum of absolute altitudinal
distances).
}
\examples{
#
}
\seealso{
.
}
