% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackToSpatialPointsDataFrame.R
\name{TrackToSpatialPointsDataFrame}
\alias{TrackToSpatialPointsDataFrame}
\title{Converts a \code{\link[trajectories]{Track}} object to a \code{\link[sp]{SpatialPointsDataFrame}}.}
\usage{
TrackToSpatialPointsDataFrame(currenttrack)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories]{Track}} object.}
}
\value{
A \code{\link[sp]{SpatialPointsDataFrame}} object
containing the information of the slot \code{sp} and \code{data}
of the corresponding \code{\link[trajectories]{Track}} object.
}
\description{
\code{TrackToSpatialPointsDataFrame} converts a
\code{\link[trajectories]{Track}} object to a
\code{\link[sp]{SpatialPointsDataFrame}}.
}
\examples{
#
}
