% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackvisits.R
\name{trackvisits}
\alias{trackvisits}
\title{Visits from a Track with Locations}
\usage{
trackvisits(
  location,
  group = NA,
  start,
  end,
  starttime,
  endtime,
  longitude,
  latitude,
  altitude,
  currentvisitduration,
  nextvisitduration = NA,
  nextvisitsamelocationindex = NA,
  nextvisitsamelocationduration = NA,
  nextvisitsamelocation = FALSE,
  campsite = FALSE,
  campsiteinbetween = FALSE,
  mergewithnext = FALSE,
  norepeatedcampsitevisits = NA
)
}
\arguments{
\item{location}{A numeric value representing the id of the locations.}

\item{group}{A numeric value representing the id of the group the
visits are assigned to.}

\item{start}{A numeric value representing a row index giving the start
of the visits.}

\item{end}{A numeric value representing a row index giving the end
of the visits.}

\item{starttime}{A \code{POSIXct} value representing the time
of the start of the visit.}

\item{endtime}{A \code{POSIXct} value representing the time
of the end of the visit.}

\item{longitude}{A numeric value representing the longitude of the
location of the visit.}

\item{latitude}{A numeric value representing the latitude of the
location of the visit.}

\item{altitude}{A numeric value representing the altitude of the
location of the visit.}

\item{currentvisitduration}{A numeric value representing the duration
between the end of the current visit and the start of the current
visit in row numbers.}

\item{nextvisitsamelocationindex}{A numeric value representing the index of
the visit that is at the same location as the current visit. The last value
must be \code{NA}.}

\item{nextvisitsamelocation}{A logical value indicating of the current visit
is classified as campsite (\code{TRUE}) or not (\code{FALSE}).}

\item{campsite}{A logical value indicating of the current visit
is classified as campsite (\code{TRUE}) or not (\code{FALSE}).}

\item{campsiteinbetween}{A logical value indicating if there is a
campsite visit at a different location between the end of the current visit
and the start of the next following visit at the same location.}

\item{mergewithnext}{A logical value indicating if the current visit
should be merged with the next visit (\code{TRUE}) or not (\code{FALSE}).
If set to \code{NULL}, \code{NA} will be inserted.}

\item{nextvisitsduration}{A numeric value representing the duration
between the end of the current visit and the start of the following
visit in row numbers. The last value must be \code{NA}.}

\item{nextvisitssamelocationduration}{A numeric value representing the duration
between the end of the current visit and the start of the next
following visit at the same location in row numbers. The last value
must be \code{NA}.}

\item{currenttrack}{A \code{data.frame} object with a variable
\code{location} that is an id for locations. Gap values must have
the value \code{0} for \code{location}.}

\item{tmin}{The minimum residence time at a specific location [s]
which is used to classify visits as campsites (if the residence
time at a specific location is larger than \code{tmin}) in
contrast to short-term visits of locations. The default is
\code{tmin = 345600}, i.e. 4 days.}

\item{tmaxinterstices}{The maximum time threshold between subsequent
visits at the same location [s]. If the duration between the two visits
is \eqn{<} \code{tmaxinterstices} and if there is no other visit
classified as campsite in-between, the two visits will be merged to one
visit.}

\item{timeinterval}{The number of seconds one data value (row) in
\code{currenttrack} covers.}
}
\value{
An object of class \code{trackvisits}. Object of class \code{trackvisits} are
\code{data.frame} objects that contain a row for each visit at a
specific location and several variables:
\describe{
  \item{\code{location}}{A numeric value representing the id of the
  location.}
  \item{\code{group}}{A numeric value representing the id of the group
  the visit is assigned to.}
  \item{\code{start}}{A numeric value representing a row index giving
  the start of the visit.}
  \item{\code{end}}{A numeric value representing a row index giving
  the end of the visit.}
  \item{\code{starttime}}{A \code{POSIXct} value representing the time
  of the start of the visit.}
  \item{\code{endtime}}{A \code{POSIXct} value representing the time
  of the end of the visit.}
  \item{\code{longitude}}{A numeric value representing the longitude of the
  location of the visit.}
  \item{\code{latitude}}{A numeric value representing the latitude of the
  location of the visit.}
  \item{\code{altitude}}{A numeric value representing the altitude of the
  location of the visit.}
  \item{\code{currentvisitduration}}{A numeric value representing the duration
  between the end of the current visit and the start of the current
  visit in row numbers.}
  \item{\code{nextvisitduration}}{A numeric value representing the duration
  between the end of the current visit and the start of the following
  visit in row numbers.}
  \item{\code{nextvisitsamelocationindex}}{A numeric value representing the index of
  the visit that is at the same location as the current visit.}
  \item{\code{nextvisitsamelocationduration}}{A numeric value representing
  the duration between the end of the current visit and the start of the next
  following visit at the same location in row numbers.}
  \item{\code{nextvisitsamelocation}}{A logical value indicating if the
  following visit is at the same location (\code{TRUE}) or not (\code{FALSE}).}
  \item{\code{campsite}}{A logical value indicating of the current visit
  is classified as campsite (\code{TRUE}) or not (\code{FALSE}).}
  \item{\code{campsiteinbetween}}{A logical value indicating if there is a
  campsite visit at a different location between the end of the current visit
  and the start of the next following visit at the same location.}
  \item{\code{mergewithnext}}{A logical value indicating if the current visit
  should be merged with the next visit (\code{TRUE}) or not (\code{FALSE}).}
  \item{\code{norepeatedcampsitevisits}}{A numeric value representing the number
  of repeated campsite visits at the same location until this visit.}
}
}
\description{
\code{trackvisits} is the constructor function of objects of class
\code{trackvisits}. Object of class \code{trackvisits} are
\code{data.frame} objects that contain a row for each visit at a
specific location and several variables for this visit:
\describe{
  \item{\code{location}}{A numeric value representing the id of the
  location.}
  \item{\code{group}}{A numeric value representing the id of the group
  the visit is assigned to.}
  \item{\code{start}}{A numeric value representing a row index giving
  the start of the visit.}
  \item{\code{end}}{A numeric value representing a row index giving
  the end of the visit.}
  \item{\code{nextvisitduration}}{A numeric value representing the duration
  between the end of the current visit and the start of the following
  visit in row numbers.}
  \item{\code{nextvisitsamelocationduration}}{A numeric value representing
  the duration between the end of the current visit and the start of the next
  following visit at the same location in row numbers.}
  \item{\code{nextvisitsamelocation}}{A logical value indicating if the
  following visit is at the same location (\code{TRUE}) or not (\code{FALSE}).}
  \item{\code{campsite}}{A logical value indicating of the current visit
  is classified as campsite (\code{TRUE}) or not (\code{FALSE}).}
  \item{\code{campsiteinbetween}}{A logical value indicating if there is a
  campsite visit at a different location between the end of the current visit
  and the start of the next following visit at the same location.}
  \item{\code{mergewithnext}}{A logical value indicating if the current visit
  should be merged with the next visit (\code{TRUE}) or not (\code{FALSE}).}
}
Object of class \code{trackvisits} can be used in order to manipulate tracks
based on visits and locations.
}
\examples{
#
}
\seealso{
\code{\link{reorganizeTracks}}, \code{\link{redefineIndices}},
\code{\link{fillGapTrack}}, \code{\link{fillGapTracks}},
\code{\link{extractClustersBuffer}}.
}
