% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackvisitsMergeGroups.R
\name{trackvisitsMergeGroups}
\alias{trackvisitsMergeGroups}
\title{Merges Grouped Visits from Track Objects}
\usage{
trackvisitsMergeGroups(currenttrackvisits, tmin, timeinterval = 30 * 60)
}
\arguments{
\item{currenttrackvisits}{An object of class \code{\link{trackvisits}} for
which groups have been defined.}
}
\value{
An object of class \code{\link{trackvisits}} in which all
visits belonging to the same group are merged into one visit. The variables
of \code{currenttrackvisits} are changed as:
\describe{
  \item{\code{location}}{not changed (location of the first visit of the group).}
  \item{\code{group}}{reset to \code{NA}.}
  \item{\code{start}}{value of \code{start} of the first visit of the group.}
  \item{\code{end}}{value of \code{end} of the last visit of the group.}
  \item{\code{starttime}}{value of \code{starttime} of the first visit of the group.}
  \item{\code{endtime}}{value of \code{endtime} of the last visit of the group.}
  \item{\code{currentvisitduration}}{updated.}
  \item{\code{nextvisitduration}}{updated.}
  \item{\code{nextvisitsamelocationduration}}{updated.}
  \item{\code{nextvisitsamelocation}}{updated.}
  \item{\code{campsite}}{updated.}
  \item{\code{campsiteinbetween}}{updated.}
  \item{\code{mergewithnext}}{reset to \code{NA}.}
}
}
\description{
\code{trackvisitsMergeGroups} merges groups of visits at the same location.
For this, an object of class \code{\link{trackvisits}} has to be processed
with \code{\link{trackvisitsGetGroups}} first or the variable
\code{trackvisits$group} has to be defined as desired.
}
\examples{
#
}
\seealso{
\code{\link{trackvisits}}.
}
