% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackvisitsSetNextvisitsamelocationindex.R
\name{trackvisitsSetNextvisitsamelocationindex}
\alias{trackvisitsSetNextvisitsamelocationindex}
\title{Check if the Next Visit is at the Same Location.}
\usage{
trackvisitsSetNextvisitsamelocationindex(currenttrackvisits)
}
\arguments{
\item{currenttrackvisits}{An object of class \code{\link{trackvisits}}.}
}
\value{
An object of class \code{\link{trackvisits}} in which
nextvisitsamelocationindex is defined.
}
\description{
\code{trackvisitsSetNextvisitsamelocationindex} sets the index of the next visit
at the same location in an object of class \code{\link{trackvisits}} is at the
same location.
}
\examples{
#
}
\seealso{
\code{\link{trackvisits}},
\code{\link{plot.trackvisits}},
\code{\link{trackvisitsFromTrack}},
\code{\link{trackvisitsGetGroups}},
\code{\link{trackvisitsMergeGroups}},
\code{\link{trackvisitsSetNorepeatedcampsitevisits}}.
}
