% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractClustersBuffer.R
\name{extractClustersBuffer}
\alias{extractClustersBuffer}
\title{Identifies clusters of points of GPS tracks.}
\usage{
extractClustersBuffer(trsSP, radius = 200)
}
\arguments{
\item{trsSP}{A \code{\link[sp]{SpatialPointsDataFrame}} object
representing a \code{\link[trajectories]{Track}} object, i.e.
the result of the function
\code{\link{TrackToSpatialPointsDataFrame}}. \code{trsSP} must
have a logical variable \code{gap} as created by
\code{\link{reorganizeTracks}}.}

\item{radius}{A numerical value representing the radius of the
buffers computed around each point [m] which are used for
clustering values to locations. Default is \code{radius = 200} [m].}
}
\value{
An integer vector with the same length as the number of points in
\code{trsSP} indicating to which cluster each point is assigned. Cluster
indices are not ordered, however points of \code{trsSP} representing gaps
(as indicated by \code{trsSP@data$gap}) get the cluster index \code{0}.
}
\description{
\code{extractClustersBuffer} identifies clusters of points in GPS
tracks based on their spatial proximity. This is done by iteratively
assigning points to buffers around selected points and merging
intersecting buffers and point sets.
}
\examples{
#
}
\seealso{
\code{\link{reorganizeTracks}}, \code{\link{redefineIndices}},
\code{\link{fillGapTrack}}, \code{\link{fillGapTracks}},
\code{\link{locationsTrack}}.
}
