% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRasterTrack.R
\name{extractRasterTrack}
\alias{extractRasterTrack}
\title{Extracts corresponding raster values for GPS tracks.}
\usage{
extractRasterTrack(
  x,
  y,
  datetime = NULL,
  method = "simple",
  buffer = 0,
  small = TRUE,
  fun = NULL,
  na.rm = TRUE,
  fixedlocationcoords = FALSE
)
}
\arguments{
\item{x}{A Raster* object. \code{x} must either have one layer or
the number of layers of \code{x} has to be the same as the number of
data values in \code{y}. If a multi-layer object is provided, it
is automatically assumed that this is the case.}

\item{y}{A \code{\link[trajectories:Track-class]{Track}} object.
Depending on other parameters set, there may be certain variables required.}

\item{datetime}{A POSIXct vector with a date (day) for each layer in \code{x}.}

\item{method}{A character value. \code{"simple"} or \code{"bilinear"}. If \code{"simple"}
values for the cell a point falls in are returned. If \code{"bilinear"} the
returned values are interpolated from the values of the four nearest raster
cells.}

\item{buffer}{A numeric value indicating the radius of the buffer
around each point that should be considered during extraction of
the raster values. If the data are not projected
(latitude/longitude), the unit should be meters. Otherwise it should
be in map-units (typically also meters).}

\item{small}{logical. If \code{TRUE} and a \code{buffer} argument is used,
the function always returns a number, also when the buffer does not include
the center of a single cell. The value of the cell in which the point falls
is returned if no cell center is within the buffer.}

\item{fun}{function to summarize the values (e.g. \code{mean}). The function
should take a single numeric vector as argument and return a single value
and accept a \code{na.rm} argument.}

\item{na.rm}{logical. Only useful when an argument fun is supplied. If
\code{na.rm = TRUE} (the default value), \code{NA} values are removed
before \code{fun} is applied. This argument may be ignored if the function
used has a \code{...} argument and ignores an additional \code{na.rm} argument.}

\item{fixedlocationcoords}{A logical value indicating if for each location in \code{y}
the same position is assumed for all data values. In this case, computation
can be speed up by setting \code{location = TRUE} and the function uses
simply the coordinates of the first value for each location.}
}
\value{
A vector with a value for each data value of \code{currenttrack}.
If a value of \code{raster} cannot be assigned to a respective data value,
\code{NA} is returned for the respective data value.
}
\description{
\code{extractRasterTrack} is the \code{\link[raster:extract]{extract}} method
in order to extract values from \code{\link[raster:raster]{Raster*}} objects
for an object of class \code{\link[trajectories:Track-class]{Track}}. Different
options are available.
}
\examples{
#
}
\seealso{
\code{\link{extractPolygonsTrack}},
\code{\link{extractRasterTracks}}.
}
