% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRasterTracks.R
\name{extractRasterTracks}
\alias{extractRasterTracks}
\title{Extracts corresponding raster values for GPS tracks.}
\usage{
extractRasterTracks(currenttracks, raster, timedate, resolution, aggregation,
  cores = 1, clcall = NULL)
}
\arguments{
\item{currenttracks}{A \code{\link[trajectories]{Tracks}} object with
\code{\link[trajectories]{Track}} objects (\code{currenttrack}) as
described in the details section.}

\item{raster}{raster based time series (\code{RasterBrick} or
\code{RasterStack} object, see: \code{\link[raster]{Raster-class}}).}

\item{timedate}{A \code{POSIXct} vector with values for each layer of
\code{raster}.}

\item{resolution}{A character value indicating the temporal resolution
of \code{raster} and \code{timedate}, respectively. One of \code{"days"},
\code{"fixedtendays"}, \code{"movingwindowtendays"} and \code{"months"}.}

\item{cores}{An integer value representing the number of cores to
use in parallel computing.}

\item{clcall}{A function that is passed to
\code{\link[parallel]{clusterCall}}.}
}
\value{
A vector with a value for each data value of \code{currenttrack}.
If a value of \code{raster} cannot be assigned to a respective data value,
\code{NA} is returned for the respective data value.
}
\description{
\code{extractRasterTracks} extracts values from raster based time series
(\code{RasterBrick} or \code{RasterStack} object, see:
\code{\link[raster]{Raster-class}}) that correspond the the respective
position and time of the \code{\link[trajectories]{Track}} objects of a
\code{\link[trajectories]{Tracks}} object.
}
\details{
The \code{Track} object have to have data values for the same time
intervals and have to be of the same length. Time values have to be
stored in a column \code{time} in the \code{data} slot for each
\code{\link[trajectories]{Track}} object.
}
\examples{
#
}
\seealso{
\code{\link{assignFixedTenDayInterval}},
\code{\link{extractRasterTracks}}.
}
