% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyBlocksVariable.R
\name{identifyBlocksVariable}
\alias{identifyBlocksVariable}
\title{Identifies data value groups within a \code{data.frame} column.}
\usage{
identifyBlocksVariable(currenttrack, variable, value)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories]{Track}} object (or
\code{data.frame} object) with a boolean column \code{gap} in
\code{currenttrack@data}. Data values have to be regularly spaced (may
be achieved for example with \code{\link{reorganizeTracks}}).}

\item{variable}{A character value indicating the name of a
variable in \code{currenttrack@data} to group for.}

\item{value}{A value of the \code{currenttrack@data$variable} for
which to identify blocks of subsequent rows.}
}
\value{
A \code{matrix} with two columns and rows for each identified
block. The first column indicates the row of \code{currenttrack@data}
where the block begins and the second column indicates the row of
\code{currenttrack@data} where the block ends or \code{NULL} if
\code{value} does not exist for \code{variable}.
}
\description{
\code{identifyBlocksVariable} identifies blocks of rows in the
\code{data} slot of a \code{\link[trajectories]{Track}} object
or \code{data.frame} object
based on a grouping variable and a value for this variable. A
block is defined as a set of directly subsequent rows with the
same value for the specified variable.
}
\examples{
#
}
\seealso{
\code{\link{reorganizeTracks}}, \code{\link{extractClutersBuffer}},
\code{\link{redefineIndices}}, \code{\link{fillGapTrack}},
\code{\link{fillGapTracks}}, \code{\link{locationsTrack}}.
}
