% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationsTrack.R
\name{locationsTrack}
\alias{locationsTrack}
\title{Identifies and classifies visits in GPS tracks.}
\usage{
locationsTrack(currenttrack, radius = 200, tmin = 345600,
  tmaxinterstices = 345600, night = c(16, 20),
  tmaxintersticenotvalid = data.frame(start =
  as.POSIXct("2016-01-01 00:00:00"), end =
  as.POSIXct("2016-05-01 00:00:00")), summary = TRUE)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories:Track-class]{Track}} object.}

\item{radius}{A numerical value representing the radius of the
buffers computed around each point [m] which are used for
clustering values to locations. Default is \code{radius = 800} [m].}

\item{tmin}{The minimum residence time at a specific location [s]
which is used to classify visits as campsites (if the residence
time at a specific location is larger than \code{tmin}) in
contrast to short-term visits of locations. The default is
\code{tmin = 345600}, i.e. 4 days.}

\item{tmaxinterstices}{The maximum time between subsequent visits
at the same location in case of which the duration of these visits
will be added in order to classify both visits together as long-term
visit (campsite) or short-term visit, based on \code{tmin}.}

\item{night}{An integer vector with two elements:
\enumerate{
  \item The first element specifies the start hour of the night, e.g. \code{0}
  for 0 o'clock.
  \item The first element specifies the start hour of the night, e.g. \code{4}
  for 4 o'clock.
}}

\item{tmaxintersticenotvalid}{A \code{data.frame} object that defines
time periods in which the argument \code{tmaxinterstice} will be ignored,
i.e. visits at the same location are merged irrespective of the duration
between these visits if there is no campsite visit at a different location
in-between. Each row indicates a time period in which this should be valid.
\code{tmaxintersticenotvalid} must contain two columns:
\describe{
  \item{\code{start}}{Represents the start time of the time interval.}
  \item{\code{end}}{Represents the end time of the time interval.}
}
It is evaluated for each visit if its endtime (\code{trackvisits$endtime})
is within any of the time periods or the starttime (\code{trackvisits$starttime})
of the next visit at the same location.}

\item{summary}{Logical value indicating if the information on the
locations and visits should be summarised (\code{summary = TRUE})
or not (\code{summary = FALSE}). See the details section for further
information.}
}
\value{
\describe{
  \item{If (\code{summary = FALSE})}{A
  \code{\link[trajectories:Track-class]{Track}} object that is identical
  to the input \code{\link[trajectories:Track-class]{Track}} object, but
  has five additional columns in the \code{data} slot:
  \describe{
    \item {\code{location}}{A numeric integer value for each identified
    spatial point cluster (location) increasing with the time starting.}
    \item {\code{visit}}{A numeric integer value for each identified
    visit increasing with the time starting.}
    \item{\code{campsite}}{A logical value indicating if a visits of a
    location is classified as long-term visit (campsite) (\code{TRUE})
    or as short-term visit (\code{FALSE}).}
    \item{\code{norepeatedcampsitevisits}}{An integer vector indicating the number of
    campsites at a specific location the data point is assigned to until the
    current visit.}
    \item{\code{start}}{A logical vector indicating if a value represents the
    first value of a visit (\code{start = TRUE}) or not (\code{start = FALSE}).}
    \item{\code{end}}{A logical vector indicating if a value represents the
    last value of a visit (\code{end = TRUE}) or not (\code{end = FALSE}).}
  }
  Gaps, as indicated by the column \code{gap}, have \code{NA} values for
  \code{location}, \code{campsite}, \code{norepeatedcampsitevisits} and \code{FALSE}
  values for \code{start} and \code{end}.
  }
  \item{If (\code{summary = TRUE})}{An object of class \code{\link{trackindicesvisits}}
  summarising the locations and visits for \code{currenttrack}.}
}
}
\description{
\code{locationsTrack} identifies clusters (locations) of
points in GPS tracks (\code{\link[trajectories:Track-class]{Track}}
object) based on their spatial proximity using
\code{\link{extractClustersBuffer}} and identifies individual
visits of the same cluster (location) along the track.
Additionally, visits are classified as long-term visits
(campsites) or short-term visits. During the identification of
clusters, data values within a user defined daily time interval
are considered. The remaining values are set to the next
location identified for the adjacent time intervals. This
procedure can be used in order to identify campsites by
assuming that a household has a campsite where it stayed
over night. If values outside the defined time interval
have a too far distance, they are classified as short-term
visit.
}
\details{
The function can be used in order to assign to each data
value of the input \code{\link[trajectories:Track-class]{Track}}
object an id of the cluster it is assigned to (
\code{summary = FALSE}) or to summarise the information
for each visit of a location (\code{summary = TRUE}).
}
\examples{
#
}
\seealso{
\code{\link{reorganizeTracks}}, \code{\link{redefineIndices}},
\code{\link{fillGapTrack}}, \code{\link{fillGapTracks}},
\code{\link{extractClustersBuffer}}.
}
