% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTracks.R
\name{mergeTracks}
\alias{mergeTracks}
\title{Merges \code{Track} objects of a \code{Tracks} object.}
\usage{
mergeTracks(currenttracks)
}
\arguments{
\item{currenttracks}{A \code{\link[trajectories:Track-class]{Tracks}} object.}
}
\value{
A  \code{\link[trajectories:Track-class]{Tracks}} object containing
the merged  \code{\link[trajectories:Track-class]{Track}} objects of the
input  \code{\link[trajectories:Track-class]{Tracks}} object as one
\code{\link[trajectories:Track-class]{Track}} object.
}
\description{
\code{mergeTracks} merges the \code{\link[trajectories:Track-class]{Track}}
objects of a \code{\link[trajectories:Track-class]{Tracks}} objects into a
singular \code{\link[trajectories:Track-class]{Tracks}} object.
}
\examples{
#
}
\seealso{
\code{\link{mergeTracksCollection}},
\code{\link{reorganizeTracks}}.
}
