% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTracksCollection.R
\name{mergeTracksCollection}
\alias{mergeTracksCollection}
\title{Merges \code{\link[trajectories]{Track}} objects of a \code{\link[trajectories]{Tracks}} objects.}
\usage{
mergeTracksCollection(trackscollection, cores = 1, clcall = NULL)
}
\arguments{
\item{currenttrackscollection}{A
\code{\link[trajectories]{TracksCollection}} object.}
}
\value{
A \code{\link[trajectories]{Tracks}} object with
a \code{\link[trajectories]{Track}} object for each former
\code{\link[trajectories]{Tracks}} object containing
the merged  \code{\link[trajectories]{Track}} objects of the
input  \code{\link[trajectories]{Tracks}} object.
}
\description{
\code{mergeTracksCollection} merges the \code{\link[trajectories]{Track}}
objects for each \code{\link[trajectories]{Tracks}} object of a
\code{\link[trajectories]{TracksCollection}} into a
singular \code{\link[trajectories]{Track}} object.
}
\examples{
#
}
\seealso{
\code{\link{mergeTracks}}, \code{\link{reorganizeTracks}}.
}
