% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointsPerQuad.R
\name{pointsPerQuad}
\alias{pointsPerQuad}
\title{Counts the number of points in the cells of a spatial grid.}
\usage{
pointsPerQuad(
  currenttrack,
  targetQsize,
  crs = "+proj=utm +zone=46 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0"
)
}
\arguments{
\item{targetQsize}{Numerical value representing the target size of the
quadrats to create.}

\item{crs}{A character string describing a projection and datum
in the \code{PROJ.4} format (see \code{\link[rgdal]{projInfo}}).}

\item{currenttracks}{A \code{\link[trajectories:Track-class]{Track}} object.}
}
\value{
A \code{\link[sp]{SpatialGridDataFrame}} with quadrat cells
and the number of points within them.
}
\description{
\code{pointsPerQuad} determines the number of points per
cell (of a given size) in a \code{\link[trajectories:Track-class]{Track}}
object.
}
\details{
This function is used by function \code{\link{extractClusters}} to
count the number of points within cells before determining clusters
of points.
}
\examples{
#
}
\seealso{
\code{\link{qTopology}}, \code{\link{extractClusters}},
\code{\link{searchNextVisit}}, \code{\link{findStarts}},
\code{\link{clusterOrder}}.
}
