% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redefineIndices.R
\name{redefineIndices}
\alias{redefineIndices}
\title{Redefines indices based on a time vector.}
\usage{
redefineIndices(x, notchange = NULL)
}
\arguments{
\item{x}{A numeric vector with integers as group indices.}

\item{notchange}{An integer vector specifying which original index values
should not be changed. Default is \code{notchange = NULL}, i.e. all indices
will be changed.}
}
\value{
A \code{data.frame} object with the redefined index values.
}
\description{
\code{redefineIndices} redefines the values of a numeric vector
containing indices (integers) as group indices so that groups
recieve values in the order their first value occurs.
}
\examples{
#
}
\seealso{
#.
}
