% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeEmptyFiles.R, R/removeEmptyFolders.R
\name{removeEmptyFiles}
\alias{removeEmptyFiles}
\title{Finds non-empty .csv files in a directory.}
\usage{
removeEmptyFiles(listname)

removeEmptyFiles(listname)
}
\arguments{
\item{listname}{A list of file names of the correpsonding GPS track files that
should be chekced.}

\item{dirlist}{A list of folder names that should be chekced.}
}
\value{
A list of file names of the correpsonding GPS track files that are not
empty.

A list of folder names that are not empty.
}
\description{
\code{removeEmptyFiles} returns the names of non-empty .csv files within a
directory by ignoring empty .csv files.

\code{removeEmptyFolders} returns the names of non-empty folders within a
directory by ignoring empty folders.
}
\examples{
#
#
}
\seealso{
\code{\link{readTrack}}, \code{\link{removeEmptyFolders}}.

\code{\link{removeEmptyFiles}}.
}
