% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryIndicatorsIntervalsTrack_distance.R
\name{summaryIndicatorsIntervalsTrack_distance}
\alias{summaryIndicatorsIntervalsTrack_distance}
\title{Computes summary indicators for GPS tracks.}
\usage{
summaryIndicatorsIntervalsTrack_distance(currenttrack, normalise = FALSE)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories:Track-class]{Track}} object that
has been processed as described in the description.}

\item{normalise}{A logical value indicating if the computed summary
indicators should be normalized relative to the duration of data values
(as specified by \code{currenttrack@data$nogapduration}) for each fixed
ten-day interval (\code{normalize = TRUE}) or not (\code{normalize = FALSE}).}
}
\value{
A numeric vector with the (normalised) distance covered within
each fixed ten-day interval \code{currenttrack} covers.
}
\description{
\code{summaryIndicatorsIntervalsTrack_distance} computes the distance
a household covered during a fixed ten-day interval. Distances are
computed for each value within a fixed ten-day interval and additionally
for the distance between the last value of the previous interval and the
first value of the target interval. The function requires a
\code{\link[trajectories:Track-class]{Track}} object as returned by
\code{\link{aggregateDailyLocationsTrack}}. Hence, only linear
distances between campsites are computed. If there is a gap with a fixed
tenday interval, it will be searched if there are values prior and after
the gap within this interval. If this is the case, the distance between
the values prior and after the gap will be summed because in this
case, it is sure that the distance has been covered within the
fixed ten-day interval. If this is not the case, the distance value of the
gap-value transition will be set to 0 which means that the will be ignored.
}
\examples{
#
}
\seealso{
\code{\link{removeDataTracks}}, \code{\link{nogapDurationTracks}}.
}
