% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryIndicatorsTrack1.R
\name{summaryIndicatorsTrack1}
\alias{summaryIndicatorsTrack1}
\title{Computes summary indicators for GPS tracks.}
\usage{
summaryIndicatorsTrack1(
  currenttrack,
  normalise = TRUE,
  timeinterval = 30 * 60,
  aggregation = "allvisits"
)
}
\arguments{
\item{currenttrack}{A \code{\link[trajectories:Track-class]{Track}} object that
has been processed as described in the description.}

\item{normalise}{A logical value indicating if the computed summary
indicators should be normalized relative to the duration of data values
(as specified by \code{currenttrack@data$nogap_duration}) for each time
interval (as specified by \code{currenttrack@data$id_timeinterval})
(\code{normalize = TRUE}) or not (\code{normalize = FALSE}).}

\item{timeinterval}{A numerical value reperesenting the duration of a
time interval represented by one data value of \code{currenttrack} [s].}

\item{aggregation}{A character value indicating for which spatial and
temporal aggregation of data values/visits at a location summary indicators
should be computed:
\describe{
  \item{\code{"raw"}}{Summary indicators will be computed on the (imputed)
  values of the original \code{\link[trajectories:Track-class]{Track}} object.}
  \item{\code{"allvisits"}}{Summary indicators will be computed on values
  aggregated for each visit, whereby it is distinguished between long-term
  visits (campsites) and short-term visits.}
  \item{\code{"campsites"}}{Like \code{"allvisits"}, but all short-term visits
  will be neglected during computations, i.e. only long-term visits will be
  considered.}
}}
}
\value{
A \code{data.frame} object with with each row representing a time
interval as specified by \code{track@data$id_timeinterval} containing the
following variables:
\describe{
  \item{\code{distance_tot}}{A numeric value representing the (normalized)
  total distance covered during a time interval [m].}
  \item{\code{altitude_tot}}{A numeric value representing the (normalized)
  total altitudinal distance covered (i.e. each metre covered in a vertical
  distance) during a time interval [m].}
  \item{\code{altitude_distance}}{A numeric value representing the (normalized)
  total altitudinal distance covered during a time interval [m] (i.e. the
  altitudinal difference between the first point in the time interval and the
  last point).}
  \item{\code{campsites_time}}{A numeric value representing the (normlaized)
  total time spend within campsites during a time interval [s].}
  \item{\code{nocampsites_time}}{A numeric value representing the (normlaized)
  total time spend not in campsites during a time interval [s] (corresponds (?)
  to the travel time).}
  \item{\code{number_campsites}}{A numeric value representing the (normalized)
  total number of unique campsites (i.e. neglecting repeated visits or counting
  locations respectively) during a time interval.}
  \item{\code{number_repeatedvisits_campsite}}{A numeric value representing the
  (normalised) total number of repeated visits for all unique campsites during a
  time interval.}
}
}
\description{
\code{summaryIndicatorsTrack1} computes various summary indicators
for a \code{\link[trajectories:Track-class]{Track}} object that has been processed
with \code{\link{reorganizeTracks}}, \code{\link{locationsTrack}} or
\code{\link{clusterOrder}}, \code{\link{removeDataTrack}} and
\code{\link{nogapDurationTrack}}.
}
\examples{
#
}
\seealso{
\code{\link{removeDataTracks}}, \code{\link{nogapDurationTracks}}.
}
